% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplex.R
\name{bioplex2}
\alias{bioplex2}
\title{Downloads the BioPlex version 2.0 interaction dataset}
\usage{
bioplex2()
}
\value{
Data frame (tibble) with interactions.
}
\description{
This dataset contains ~56,000 interactions detected in HEK293T cells
using 5,891 baits.
More details at https://bioplex.hms.harvard.edu/interactions.php
}
\examples{
bioplex_interactions <- bioplex2()
nrow(bioplex_interactions)
# [1] 56553
colnames(bioplex_interactions)
# [1] "GeneA"         "GeneB"        "UniprotA"   "UniprotB"
# [5] "SymbolA"       "SymbolB"      "p_wrong"    "p_no_interaction"
# [9] "p_interaction"

}
\seealso{
\itemize{
    \item{\code{\link{bioplex1}}}
    \item{\code{\link{bioplex3}}}
    \item{\code{\link{bioplex_hct116_1}}}
    \item{\code{\link{bioplex_all}}}
}
}
