% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplex.R
\name{bioplex3}
\alias{bioplex3}
\title{Downloads the BioPlex version 3.0 interaction dataset}
\usage{
bioplex3()
}
\value{
Data frame (tibble) with interactions.
}
\description{
This dataset contains ~120,000 interactions detected in HEK293T cells
using 10,128 baits.
More details at \url{https://bioplex.hms.harvard.edu/interactions.php}.
}
\examples{
bioplex_interactions <- bioplex3()
nrow(bioplex_interactions)
# [1] 118162
colnames(bioplex_interactions)
# [1] "GeneA"         "GeneB"        "UniprotA"   "UniprotB"
# [5] "SymbolA"       "SymbolB"      "p_wrong"    "p_no_interaction"
# [9] "p_interaction"

}
\seealso{
\itemize{
    \item{\code{\link{bioplex1}}}
    \item{\code{\link{bioplex2}}}
    \item{\code{\link{bioplex_hct116_1}}}
    \item{\code{\link{bioplex_all}}}
}
}
