% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma_motif_es}
\alias{bma_motif_es}
\title{Prints a BMA motif to the screen from a sequence of edges, which can be
copy/pasted into the BMA canvas}
\usage{
bma_motif_es(edge_seq, G, granularity = 2)
}
\arguments{
\item{edge_seq}{An igraph edge sequence.}

\item{G}{An igraph graph object.}

\item{granularity}{Numeric: granularity value.}
}
\description{
Intended to parallel print_path_es
}
