% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma_motif_vs}
\alias{bma_motif_vs}
\title{Prints a BMA motif to the screen from a sequence of nodes, which can be
copy/pasted into the BMA canvas}
\usage{
bma_motif_vs(node_seq, G)
}
\arguments{
\item{node_seq}{An igraph node sequence.}

\item{G}{An igraph graph object.}
}
\description{
Intended to parallel print_path_vs
}
