% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chalmers_gem.R
\name{chalmers_gem}
\alias{chalmers_gem}
\title{Alberto Valdeolivas
Genome scale metabolic model by Wang et al. 2021}
\usage{
chalmers_gem(organism = "Human", orphans = TRUE)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier.
Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus),
10116 (Rattus norvegicus), 7955 (Danio rerio), 7227 (Drosophila
melanogaster) and 6239 (Caenorhabditis elegans).}

\item{orphans}{Logical: include orphan reactions (reactions without known
enzyme).}
}
\value{
List containing the following elements: \itemize{
    \item{reactions: tibble of reaction data;}
    \item{metabolites: tibble of metabolite data;}
    \item{reaction_ids: translation table of reaction identifiers;}
    \item{metabolite_ids: translation table of metabolite identifiers;}
    \item{S: Stoichiometric matrix (sparse).}
}
}
\description{
Process the GEMs from Wang et al., 2021
(\url{https://github.com/SysBioChalmers}) into convenient tables.
}
\examples{
gem <- chalmers_gem()

}
\references{
Wang H, Robinson JL, Kocabas P, Gustafsson J, Anton M, Cholley
    PE, Huang S, Gobom J, Svensson T, Uhlen M, Zetterberg H, Nielsen J.
    Genome-scale metabolic network reconstruction of model animals as a
    platform for translational research. Proc Natl Acad Sci U S A. 2021 Jul
    27;118(30):e2102344118. doi: \doi{10.1073/pnas.2102344118}.
}
