% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{common_name}
\alias{common_name}
\title{Common (English) names of organisms}
\usage{
common_name(name)
}
\arguments{
\item{name}{Vector with any kind of organism name or identifier, can be
also mixed type.}
}
\value{
Character vector with common (English) taxon names, NA if a name
    in the input could not be found.
}
\description{
Common (English) names of organisms
}
\examples{
common_name(c(10090, "cjacchus", "Vicugna pacos"))
# [1] "Mouse" "White-tufted-ear marmoset" "Alpaca"

}
