% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuspathdb.R
\name{consensuspathdb_raw_table}
\alias{consensuspathdb_raw_table}
\title{Downloads interaction data from ConsensusPathDB}
\usage{
consensuspathdb_raw_table()
}
\description{
Downloads interaction data from ConsensusPathDB
}
\examples{
cpdb_raw <- consensuspathdb_raw_table()

}
