% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmos.R
\name{cosmos_pkn}
\alias{cosmos_pkn}
\title{Alberto Valdeolivas
Prior knowledge network (PKN) for COSMOS}
\usage{
cosmos_pkn(
  organism = "human",
  protein_ids = c("uniprot", "genesymbol"),
  metabolite_ids = c("hmdb", "kegg"),
  chalmers_gem_metab_max_degree = 400L,
  stitch_score = 700L,
  ...
)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier.
Supported taxons are 9606 (\emph{Homo sapiens}), 10090
(\emph{Mus musculus}), and 10116 (\emph{Rattus norvegicus}).}

\item{protein_ids}{Character: translate the protein identifiers to these ID
types. Each ID type results two extra columns in the output, for the "a"
and "b" sides of the interaction, respectively. The default ID type for
proteins is Esembl Gene ID, and by default UniProt IDs and Gene
Symbols are included.}

\item{metabolite_ids}{Character: translate the protein identifiers to these ID
types. Each ID type results two extra columns in the output, for the "a"
and "b" sides of the interaction, respectively. The default ID type for
metabolites is Metabolic Atlas ID, and HMDB IDs and KEGG IDs are included.}

\item{chalmers_gem_metab_max_degree}{Degree cutoff used to filter out
metabolites (400 by default). The objective is to remove cofactors and
over-promiscuous metabolites.}

\item{stitch_score}{Confidence cutoff used for STITCH connections (700 by
default).}

\item{...}{Further parameters to \code{\link{import_omnipath_interactions}}.}
}
\value{
List of 4 elements containing the necessary information for COSMOS to
    run: causal PKN, mapping data frame for metabolites from GEM,
    reaction-to-gene data frame from GEM, and mapping data frame for reactions
    from gem_
}
\description{
This function generates the prior knowledge network (PKN) needed to run
COSMOS using information from different resources through the \pkg{OmnipathR}
R package. Particularly, \code{cosmos_pkn} will obtain:
\itemize{ \item Genome-scale metabolic
model (GEM) of the required organism from Wang et al., 2021.
\item Interaction network of
chemicals and proteins from STITCH (\url{http://stitch.embl.de/}) for the
required organism. \item Protein-protein interactions from Omnipath (Türei
et al., 2021) for the required organism} With these three pieces of
information, the function will generate the required causal network for
COSMOS to run.
}
\examples{
\dontrun{
    human_cosmos <- cosmos_pkn(organism = 9606)
}

}
\references{
Wang H, Robinson JL, Kocabas P, Gustafsson J, Anton M,
       Cholley PE, et al. Genome-scale metabolic network reconstruction of model
       animals as a platform for translational research. Proceedings of the
       National Academy of Sciences. 2021 Jul 27;118(30):e2102344118.

       Türei D, Valdeolivas A, Gul L, Palacio‐Escat N, Klein M, Ivanova O, et al.
        Integrated intra‐ and intercellular signaling knowledge for multicellular
        omics analysis. Molecular Systems Biology. 2021 Mar;17(3):e9923.
}
