% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmos_PKN_generation.R
\name{create_PKN_COSMOS}
\alias{create_PKN_COSMOS}
\title{Generating COSMOS' PKN for different organisms}
\usage{
create_PKN_COSMOS(
  organism,
  translate.genes = FALSE,
  biomart.use.omnipath = TRUE,
  GEM.reactions.map.col = "rxns",
  GEM.metabolites.map.col = "mets",
  GEM.list.params = list(stoich.name = "S", reaction.name = "grRules", lb.name = "lb",
    ub.name = "ub", rev.name = "rev", reaction.ID.name = "rxns", metabolites.ID.name =
    "mets", metabolites.names.name = "metNames", metabolites.fomulas.name =
    "metFormulas", metabolites.inchi.name = "inchis"),
  GEM.degree.mets.threshold = 400,
  stitch.threshold = 700,
  verbose = TRUE
)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier. 
Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus), 
10116 (Rattus norvegicu), 7955 (Danio rerio), 7227 (Drosophila 
melanogaster) and 6239 (Caenorhabditis elegans).}

\item{translate.genes}{Whether translating genes from ENSEMBL into SYMBOL. 
\code{FALSE} by default.}

\item{biomart.use.omnipath}{Whether using BiomaRt information from OmnipathR 
(\code{TRUE} by default).}

\item{GEM.reactions.map.col}{Column of reaction IDs in the GEM
(\code{"rxns"} by default).}

\item{GEM.metabolites.map.col}{Column of reaction IDs in the GEM 
(\code{"mets"} by default).}

\item{GEM.list.params}{List containing the name of the slots where the 
information to construct the PKN is located in the GEM. If a matlab object
is provided, this list parameter should not be modified.}

\item{GEM.degree.mets.threshold}{Degree cutoff used to filter out 
metabolites (400 by default). The objective is to remove cofactors and 
other metabolites with many connections.}

\item{stitch.threshold}{Confidence cutoff used for STITCH connections 
(700 by default).}

\item{verbose}{Whether showing messages during execution (\code{TRUE} by 
default).}
}
\value{
List of 4 elements containing the necessary information for COSMOS to 
  run: causal PKN,
}
\description{
This function generates the prior knowledge network (PKN) needed to run 
COSMOS using information from different resources. It will download the 
required information through the \pkg{OmnipathR} R package. Particularly, 
\code{create_PKN_COSMOS} will obtain: \itemize{ \item Genome-scale metabolic 
model (GEM) of the required organism from (). \item Interaction network of 
chemical and proteins from STITCH (\url(http://stitch.embl.de/)) for the 
required organism. \item Protein-protein interactions from Omnipath ()} 
With these three pieces of information, the function will generate the 
required causal network for COSMOS to run.
}
\examples{
\dontrun{
  human.PKN.COSMOS <- create_PKN_COSMOS(organism = 9606)
}

}
