% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{database_summary}
\alias{database_summary}
\title{Summary of the annotations and intercell database contents}
\usage{
database_summary(query_type, return_df = FALSE)
}
\arguments{
\item{query_type}{Character: either "annotations" or "intercell".}

\item{return_df}{Logical: return a data frame instead of list.}
}
\value{
Summary of the database contents: the available resources, fields,
    and their possible values. As a nested list if format is "json",
    otherwise a data frame.
}
\description{
The `annotations_summary` and `intercell_summary` query types return
detailed information on the contents of these databases. It includes
all the available resources, fields and values in the database.
}
\examples{
database_summary('annotations')

}
