% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{datasets_one_column}
\alias{datasets_one_column}
\title{Create a column with dataset names listed}
\usage{
datasets_one_column(data, remove_logicals = TRUE)
}
\arguments{
\item{data}{Interactions data frame with dataset columns (i.e. queried
with the option `fields = "datasets"`).}

\item{remove_logicals}{Logical: remove the per dataset logical columns.}
}
\value{
The input data frame with the new column "datasets" added.
}
\description{
From logical columns for each dataset, here we create a column that is
a list of character vectors, containing dataset labels.
}
