% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{.get_annotation_databases}
\alias{.get_annotation_databases}
\title{Get the different annotation databases integrated in Omnipath}
\usage{
.get_annotation_databases()
}
\value{
character vector with the names of the annotation databases
}
\description{
get the names of the databases from \url{http://omnipath.org/annotation}
}
\examples{
.get_annotation_databases()
}
\seealso{
\code{\link{import_Omnipath_annotations}}
}
