% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{.get_interaction_databases}
\alias{.get_interaction_databases}
\title{Get interaction databases}
\usage{
.get_interaction_databases()
}
\value{
character vector with the databases
}
\description{
get the names of the databases from omnipath.org/interactions
}
