% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{.get_ptms_databases}
\alias{.get_ptms_databases}
\title{Get ptms databases}
\usage{
.get_ptms_databases()
}
\value{
character vector with the databases
}
\description{
get the names of the databases from omnipath.org/ptms
}
