% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.slow_doctest}
\alias{.slow_doctest}
\title{Bypass slow doctests}
\usage{
.slow_doctest(value = NULL, env = parent.frame())
}
\arguments{
\item{value}{The value to return.}
}
\value{
This function does not return anything, its parent returns
    `value`.
}
\description{
Makes the calling function return a value, bypassing further execution.
Used in functions that normally take a long time to run, and hence, due
to the 40 minutes timeout on Bioconductor build servers, we can not afford
to run them. We still run these doctests daily on our own build server,
so they fulfill their testing purpose.
}
