% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl_dataset}
\alias{ensembl_dataset}
\title{Ensembl dataset name from organism}
\usage{
ensembl_dataset(organism)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) name or
identifier. If an Ensembl dataset name is provided}
}
\value{
Character: name of an ensembl dataset.
}
\description{
Ensembl dataset name from organism
}
\examples{
ensembl_dataset(10090)
# [1] "mmusculus_gene_ensembl"

}
