% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{ensembl_id_type}
\alias{ensembl_id_type}
\title{Ensembl identifier type label}
\usage{
ensembl_id_type(label)
}
\arguments{
\item{label}{Character: an ID type label, as shown in the table at
\code{\link{translate_ids}}}
}
\value{
Character: the Ensembl specific ID type label, or the input
    unchanged if it could not be translated (still might be a valid
    identifier name). These labels should be valid Ensembl attribute
    names, directly usable in Ensembl queries.
}
\description{
Ensembl identifier type label
}
\examples{
ensembl_id_type("uniprot")
# [1] "uniprotswissprot"

}
\seealso{
\itemize{
    \item{\code{\link{uniprot_id_type}}}
    \item{\code{\link{uploadlists_id_type}}}
}
}
