% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl_organisms}
\alias{ensembl_organisms}
\title{Organism names and identifiers from Ensembl}
\usage{
ensembl_organisms()
}
\value{
A data frame with the above mentioned columns.
}
\description{
A table with various taxon names and identifiers: English common names,
latin (scientific) names, Ensembl organism IDs and NCBI taxonomy IDs.
}
\examples{
ens_org <- ensembl_organisms()
ens_org

}
