% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl_organisms_raw}
\alias{ensembl_organisms_raw}
\title{Table of Ensembl organisms}
\usage{
ensembl_organisms_raw()
}
\value{
The table described above as a data frame.
}
\description{
A table with various taxon IDs and metadata about related Ensembl database
contents, as shown at https://www.ensembl.org/info/about/species.html.
The "Taxon ID" column contains the NCBI Taxonomy identifiers.
}
\examples{
ens_org <- ensembl_organisms_raw()
ens_org

}
