% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ensure_igraph}
\alias{ensure_igraph}
\title{Converts a network to igraph object unless it is already one}
\usage{
ensure_igraph(network)
}
\arguments{
\item{network}{Either an OmniPath interaction data frame, or an igraph
graph object.}
}
\value{
An igraph graph object.
}
\description{
Converts a network to igraph object unless it is already one
}
