% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{extra_attr_to_column}
\alias{extra_attr_to_column}
\title{New column from one extra attribute}
\usage{
extra_attr_to_column(data, attr, flatten = FALSE)
}
\arguments{
\item{data}{An interaction data frame.}

\item{attr}{The name of an extra attribute; NSE is supported.}

\item{flatten}{Logical: unnest the list column even if some records have
multiple values for the attributes; these will yield multiple records
in the resulted data frame.}
}
\value{
Data frame with the new column created; the new column is list
    type if one interaction might have multiple values of the attribute,
    or character type if
}
\description{
New column from one extra attribute
}
