% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{filter_by_resource}
\alias{filter_by_resource}
\title{Filters OmniPath data by resources}
\usage{
filter_by_resource(data, resources = NULL)
}
\arguments{
\item{data}{A data frame downloaded from the OmniPath web service
(interactions, enzyme-substrate or complexes).}

\item{resources}{Character vector with resource names to keep.}
}
\value{
The data frame filtered.
}
\description{
Keeps only those records which are supported by any of the resources of
interest.
}
\examples{
interactions <- import_omnipath_interactions()
signor <- filter_by_resource(interactions, resources = 'SIGNOR')

}
