% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidences.R
\name{filter_evidences}
\alias{filter_evidences}
\title{Filter evidences by dataset, resource and license}
\usage{
filter_evidences(data, ..., datasets = NULL, resources = NULL, exclude = NULL)
}
\arguments{
\item{data}{An interaction data frame with some columns containing evidences
as nested lists.}

\item{...}{The evidences columns to filter: tidyselect syntax is supported.
By default the columns "evidences", "positive", "negative", "directed"
and "undirected" are filtered, if present.}

\item{datasets}{A character vector of dataset names.}

\item{resources}{A character vector of resource names.}

\item{exclude}{Character vector of resource names to be excluded.}
}
\value{
The input data frame with the evidences in the selected columns
   filtered.
}
\description{
Filter evidences by dataset, resource and license
}
\seealso{
\itemize{
    \item{\code{\link{only_from}}}
    \item{\code{\link{unnest_evidences}}}
    \item{\code{\link{from_evidences}}}
}
}
