% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{get_annotation_resources}
\alias{get_annotation_resources}
\alias{get_annotation_databases}
\title{Retrieves a list of available resources in the annotations database
of OmniPath}
\usage{
get_annotation_resources(dataset = NULL, ...)
}
\arguments{
\item{dataset}{ignored for this query type}

\item{...}{optional additional arguments}
}
\value{
character vector with the names of the annotation resources
}
\description{
Get the names of the resources from \url{https://omnipath.org/annotations}.
}
\examples{
get_annotation_resources()

}
\seealso{
\itemize{
    \item{\code{\link{get_resources}}}
    \item{\code{\link{import_omnipath_annotations}}}
}
}
