% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complex_genes.R
\name{get_complex_genes}
\alias{get_complex_genes}
\title{Get all the molecular complexes for a given gene(s)}
\usage{
get_complex_genes(complexes = import_Omnipath_complexes(),
  select_genes = query_genes, total_match = FALSE)
}
\arguments{
\item{complexes}{complexes data frame (obtained using 
\code{\link{import_Omnipath_complexes}})}

\item{select_genes}{vector containing the genes for whom complexes will be
retrieved (hgnc format).}

\item{total_match}{[default=FALSE] logical indicating if the user wants to get all 
the complexes where any of the input genes participate (FALSE) or
to get only the complexes where all the input genes participate together 
(TRUE)}
}
\value{
data.frame of complexes
}
\description{
This function returns all the molecular complexes where an input set
of genes participate. User can choose to retrieve every complex where 
any of the input genes participate or just retrieve these complexes where
all the genes in input set participate together.
}
\examples{
complexes = import_Omnipath_complexes(filter_databases=c("CORUM", "hu.MAP"))
query_genes = c("LMNA","BANF1")
complexes_query_genes = get_complex_genes(complexes,query_genes)
}
\seealso{
\code{\link{import_Omnipath_complexes}})
}
