% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{get_complex_resources}
\alias{get_complex_resources}
\alias{get_complexes_databases}
\title{Retrieve a list of complex resources available in Omnipath}
\usage{
get_complex_resources(dataset = NULL)
}
\arguments{
\item{dataset}{ignored for this query type}
}
\value{
character vector with the names of the databases
}
\description{
get the names of the resources from \url{https://omnipath.org/complexes}
}
\examples{
\donttest{
get_complex_resources()
}

}
\seealso{
\code{\link{get_resources},
\link{import_omnipath_complexes}}
}
