% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_complexes_databases}
\alias{get_complexes_databases}
\title{Get the different complexes databases integrated in Omnipath}
\usage{
get_complexes_databases()
}
\value{
character vector with the names of the databases
}
\description{
get the names of the databases from \url{http://omnipath.org/complexes}
}
\examples{
get_complexes_databases()
}
\seealso{
\code{\link{import_Omnipath_complexes}}
}
