% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_intercell_categories}
\alias{get_intercell_categories}
\title{Retrieve a list of categories from the intercell database of OmniPath}
\usage{
get_intercell_categories()
}
\value{
character vector with the different intercell categories
}
\description{
get the names of the categories from \url{http://omnipath.org/intercell}
}
\examples{
get_intercell_categories()
}
\seealso{
\code{\link{import_omnipath_intercell},
\link{get_intercell_classes}}
}
