% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_intercell_classes}
\alias{get_intercell_classes}
\title{Get the different intercell main classes described in Omnipath}
\usage{
get_intercell_classes()
}
\value{
character vector with the different intercell main classes
}
\description{
get the names of the main classes from \url{http://omnipath.org/intercell}
}
\examples{
get_intercell_classes()
}
\seealso{
\code{\link{import_Omnipath_intercell},
\link{get_intercell_categories}}
}
