% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_intercell_resources}
\alias{get_intercell_resources}
\title{Retrieve a list of intercellular communication resources available in
Omnipath}
\usage{
get_intercell_resources(dataset = NULL)
}
\arguments{
\item{dataset}{ignored at this query type}
}
\value{
character vector with the names of the databases
}
\description{
get the names of the databases from \url{http://omnipath.org/intercell}
}
\examples{
get_intercell_resources()

}
\seealso{
\code{\link{get_resources},
\link{import_omnipath_intercell}}
}
