% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signed_ptms.R
\name{get_signed_ptms}
\alias{get_signed_ptms}
\title{get signs for ptms interactions}
\usage{
get_signed_ptms(
  ptms = import_Omnipath_PTMS(),
  interactions = import_Omnipath_Interactions()
)
}
\arguments{
\item{ptms}{ptms data frame generated by \code{\link{import_Omnipath_PTMS}}}

\item{interactions}{interaction data frame generated by
\code{\link{import_Omnipath_Interactions}}}
}
\value{
data.frame of ptms with is_inhibition and is_stimulation columns
}
\description{
ptms data does not contain sign (activation/inhibition), we generate this
information based on the interaction network
}
\examples{
ptms = import_Omnipath_PTMS(filter_databases=c("PhosphoSite", "SIGNOR"))
interactions = import_Omnipath_Interactions()
ptms = get_signed_ptms(ptms,interactions)
}
\seealso{
\code{\link{import_Omnipath_PTMS}}
\code{\link{import_Omnipath_Interactions}}
}
