% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{giant_component}
\alias{giant_component}
\title{Giant component of a graph}
\usage{
giant_component(graph)
}
\arguments{
\item{graph}{An igraph graph object.}
}
\value{
An igraph graph object containing only the giant component.
}
\description{
For an igraph graph object returns its giant component.
}
\examples{
interactions <- import_post_translational_interactions()
graph <- interaction_graph(interactions)
graph_gc <- giant_component(graph)

}
