% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go_ontology_download}
\alias{go_ontology_download}
\title{Downloads and preprocesses the Gene Ontology tree}
\usage{
go_ontology_download(
  relations = c("is_a", "part_of", "occurs_in", "regulates", "positively_regulates",
    "negatively_regulates"),
  go_variant = "go-basic"
)
}
\description{
Downloads and preprocesses the Gene Ontology tree
}
