% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_interaction}
\alias{graph_interaction}
\title{Interaction data frame from igraph graph object}
\usage{
graph_interaction(graph, implode = FALSE)
}
\arguments{
\item{graph}{An igraph graph object created formerly from an OmniPath
interactions data frame.}

\item{implode}{Logical: restore the original state of the list type
columns by imploding them to character vectors, subitems separated
by semicolons.}
}
\value{
An interaction data frame.
}
\description{
Convert an igraph graph object to interaction data frame. This is the
reverse of the operation done by thje \code{\link{interaction_graph}}
function. Networks can be easily converted to igraph objects, then
you can make use of all igaph methods, and at the end, get back the
interactions in a data frame, along with all new edge and node attributes.
}
\seealso{
\code{\link{interaction_graph}}
}
