% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_attrs.R
\name{has_extra_attrs}
\alias{has_extra_attrs}
\title{Tells if an interaction data frame has an extra_attrs column}
\usage{
has_extra_attrs(data)
}
\arguments{
\item{data}{An interaction data frame.}
}
\value{
Logical: TRUE if the data frame has the "extra_attrs" column.
}
\description{
Tells if an interaction data frame has an extra_attrs column
}
\examples{
i <- import_omnipath_interactions(fields = 'extra_attrs')
has_extra_attrs(i)

}
\seealso{
\itemize{
    \item{\code{\link{extra_attrs}}}
    \item{\code{\link{extra_attrs_to_cols}}}
    \item{\code{\link{with_extra_attrs}}}
}
}
