% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{hmdb_id_type}
\alias{hmdb_id_type}
\title{HMDB identifier type label}
\usage{
hmdb_id_type(label)
}
\arguments{
\item{label}{Character: an ID type label, as shown in the table at
\code{\link{translate_ids}}}
}
\value{
Character: the HMDB specific ID type label, or the input
    unchanged if it could not be translated (still might be a valid
    identifier name). These labels should be valid HMDB field
    names, as used in HMDB XML files.
}
\description{
HMDB identifier type label
}
\examples{
hmdb_id_type("hmdb")
# [1] "accession"

}
\seealso{
\itemize{
    \item{\code{\link{uniprot_id_type}}}
    \item{\code{\link{ensembl_id_type}}}
    \item{\code{\link{uploadlists_id_type}}}
}
}
