% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdb.R
\name{hmdb_protein_fields}
\alias{hmdb_protein_fields}
\title{Field names for the HMDB proteins dataset}
\usage{
hmdb_protein_fields()
}
\value{
Character vector of field names.
}
\description{
Field names for the HMDB proteins dataset
}
\examples{
hmdb_protein_fields()

}
\seealso{
\itemize{
    \item{\code{\link{hmdb_table}}}
    \item{\code{\link{hmdb_metabolite_fields}}}
}
}
