% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologene.R
\name{homologene_organisms}
\alias{homologene_organisms}
\title{Organisms in NCBI HomoloGene}
\usage{
homologene_organisms(name_type = "ncbi")
}
\arguments{
\item{name_type}{Character: type of the returned name or identifier.
Many synonyms are accepted, the shortest ones: "latin", "ncbi",
"common", "ensembl". Case unsensitive.}
}
\value{
A character vector of organism names.
}
\description{
Organisms in NCBI HomoloGene
}
\details{
Not all NCBI Taxonomy IDs can be translated to common or
    latin names. It means some organisms will be missing if translated
    to those name types. In the future we will address this issue, until
    then if you want to see all organisms use NCBI Taxonomy IDs.
}
