% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{id_translation_resources}
\alias{id_translation_resources}
\title{List available ID translation resources}
\usage{
id_translation_resources()
}
\value{
A character vector with the names of the available ID translation
    resources.
}
\description{
List available ID translation resources
}
\examples{
id_translation_resources()

}
