% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{id_types}
\alias{id_types}
\title{ID types and synonyms in identifier translation}
\usage{
id_types()
}
\value{
Data frame with 4 columns: the ID type labels in the resource, their
    synonyms in OmniPath (this package), the name of the ID translation
    resource, and the entity type.
}
\description{
ID types and synonyms in identifier translation
}
\examples{
id_types()

}
\seealso{
\itemize{
    \item{\code{\link{translate_ids}}}
    \item{\code{\link{translate_ids_multi}}}
    \item{\code{\link{ensembl_id_mapping_table}}}
    \item{\code{\link{uniprot_id_mapping_table}}}
    \item{\code{\link{hmdb_id_mapping_table}}}
    \item{\code{\link{chalmers_gem_id_mapping_table}}}
    \item{\code{\link{uniprot_full_id_mapping_table}}}
    \item{\code{\link{ensembl_id_type}}}
    \item{\code{\link{uniprot_id_type}}}
    \item{\code{\link{hmdb_id_type}}}
    \item{\code{\link{chalmers_gem_id_type}}}
}
}
