% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{import_all_interactions}
\alias{import_all_interactions}
\alias{import_AllInteractions}
\title{Imports all interaction datasets available in OmniPath}
\usage{
import_all_interactions(
  resources = NULL,
  organism = 9606,
  dorothea_levels = c("A", "B"),
  exclude = NULL,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  strict_evidences = FALSE,
  types = NULL,
  ...
)

import_AllInteractions(...)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{dorothea_levels}{The confidence levels of the dorothea
interactions (TF-target) which range from A to D. Set to A and B by
default.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{types}{Character: interaction types, such as "transcriptional",
"post_transcriptional", "post_translational", etc.}

\item{...}{Passed to \code{import_all_interactions}.}
}
\value{
A dataframe containing all the datasets in the interactions query
}
\description{
The interaction datasets currently available in OmniPath:
}
\details{
omnipath: the OmniPath data as defined in the paper, an arbitrary optimum
between coverage and quality
pathwayextra: activity flow interactions without literature reference
kinaseextra: enzyme-substrate interactions without literature reference
ligrecextra: ligand-receptor interactions without literature reference
dorothea: transcription factor (TF)-target interactions from DoRothEA
tf_target: transcription factor (TF)-target interactions from other
resources
mirnatarget: miRNA-mRNA interactions
tf_mirna: TF-miRNA interactions
lncrna_mrna: lncRNA-mRNA interactions
}
\examples{
interactions <- import_all_interactions(
    resources = c('HPRD', 'BioGRID'),
    organism = 9606
)

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
