% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_AllInteractions}
\alias{import_AllInteractions}
\title{Imports from Omnipath webservice all the available interactions 
from the different datasets}
\usage{
import_AllInteractions(from_cache_file = NULL,
  filter_databases = .get_interaction_databases(),
  select_organism = 9606)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are 
removed. See \code{\link{.get_interaction_databases}} for more information.}

\item{select_organism}{Interactions are available for human, mouse and rat. 
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}
}
\value{
A dataframe containing all the datasets in the interactions query
}
\description{
Imports the dataset from: 
\url{http://omnipathdb.org/interactions?datasets=omnipath,pathwayextra,
kinaseextra,ligrecextra,tfregulons,mirnatarget&fields=sources,
references&genesymbols=1}, 
which contains all the different interactions available in the webserver:
}
\details{
omnipath: the OmniPath data as defined in the paper, an arbitrary optimum 
between coverage and quality
pathwayextra: activity flow interactions without literature reference
kinaseextra: enzyme-substrate interactions without literature reference
ligrecextra: ligand-receptor interactions without literature reference
tfregulons: transcription factor (TF)-target interactions from DoRothEA
mirnatarget: miRNA-mRNA and TF-miRNA interactions
}
\examples{
interactions <- import_AllInteractions(filter_databases=c("HPRD","BioGRID"),
    select_organism = 9606)
}
\seealso{
\code{\link{.get_interaction_databases}}
}
