% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_dorothea_interactions}
\alias{import_dorothea_interactions}
\alias{import_TFregulons_Interactions}
\alias{import_tfregulons_interactions}
\title{From the OmniPath webservice imports interactions from the
DoRothEA dataset}
\usage{
import_dorothea_interactions(
  resources = NULL,
  organism = 9606,
  dorothea_levels = c("A", "B"),
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{dorothea_levels}{Vector detailing the confidence levels of the
interactions to be downloaded. In dorothea, every TF-target interaction
has a confidence score ranging from A to E, being A the most reliable
interactions.
By default we take A and B level interactions (\code{c(A, B)}).
It is to note that E interactions are not available in OmnipathR.}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing TF-target interactions from DoRothEA
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=dorothea}
which contains transcription factor (TF)-target interactions from DoRothEA
\url{https://github.com/saezlab/DoRothEA}
}
\examples{
interactions <- import_dorothea_interactions(
    resources = c('DoRothEA', 'ARACNe-GTEx_DoRothEA'),
    organism = 9606,
    dorothea_levels = c('A', 'B', 'C')
)

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
}
}
