% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_intercell_network}
\alias{import_intercell_network}
\title{Imports an intercellular network combining annotations and interactions}
\usage{
import_intercell_network(
  from_cache_file = NULL,
  filter_databases = get_interaction_databases(),
  classes_source = list(transmiters = c("ligand"), receivers = c("receptor"))
)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{vector containing interactions databases. 
Interactions not reported in these databases are removed. 
See \code{\link{get_interaction_databases}} for more information.}

\item{classes_source}{A list containing two vectors. The first one with 
the main classes to be considered as transmiters and the second with the 
main classes to be considered as receivers. For furter information
about the main classes see \code{\link{get_intercell_classes}}}
}
\value{
A dataframe containing information about protein-protein 
interactions and the inter-cellular roles of the protiens involved in those
interactions.
}
\description{
Imports an intercellular network by mapping intercellular annotations 
and protein interactions. It first imports the PPI interactions from the
different datasets here described. Then, it takes proteins with the 
intercellular roles defined by the user. Some proteins should be defined 
as transmiters (eg. ligand) and other as receivers (receptor). We find the 
interactions which source is a transmiter and its target a receiver.
}
\examples{
intercellNetwork <- import_intercell_network(
classes_source = list(transmiters=c('ligand'),receivers=c('receptor')))
}
\seealso{
\code{\link{get_intercell_categories}}
}
