% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{import_ligrecextra_interactions}
\alias{import_ligrecextra_interactions}
\title{Imports interactions from the `ligrec extra` dataset of OmniPath}
\usage{
import_ligrecextra_interactions(
  resources = NULL,
  organism = "human",
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Character or integer: Name or NCBI Taxonomy ID of one or
organisms. The web service currently provides interactions for
human, mouse and rat. For other organisms, the data will be translated
by orthologous gene pairs from human. In this case, only one organism
can be provided.}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing ligand-receptor interactions including
the ones without literature references
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=ligrecextra},
which contains ligand-receptor interactions without literature reference.
The ligand-receptor interactions supported by literature references
are part of the `omnipath` dataset.
}
\examples{
interactions <- import_ligrecextra_interactions(
    resources = c('HPRD', 'Guide2Pharma'),
    organism = 9606
)

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
