% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_lncrna_mrna_interactions}
\alias{import_lncrna_mrna_interactions}
\title{Imports interactions from the lncRNA-mRNA dataset of OmniPath}
\usage{
import_lncrna_mrna_interactions(
  cache_file = NULL,
  resources = NULL,
  organism = 9606,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  ...
)
}
\arguments{
\item{cache_file}{path to an earlier data file}

\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_databases}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{fields}{The user can define here the fields to be added. If used, set 
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing lncRNA-mRNA interactions
}
\description{
Imports the dataset from:
\url{http://omnipathdb.org/interactions?datasets=lncrna_mrna},
which contains lncRNA-mRNA interactions
}
\examples{
interactions <-
    import_lncrna_mrna_interactions(
        resources = c('ncRDeathDB')
    )

}
\seealso{
\code{\link{get_interaction_databases},
  \link{import_all_interactions}}
}
