% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_Omnipath_annotations}
\alias{import_Omnipath_annotations}
\title{Import Omnipath Annotations}
\usage{
import_Omnipath_annotations(
  from_cache_file = NULL,
  select_genes = NULL,
  filter_databases = NULL,
  force_full_download = FALSE,
  ...
)
}
\arguments{
\item{from_cache_file}{Path to an earlier data file}

\item{select_genes}{Vector containing the genes or proteins for whom
annotations will be retrieved (UniProt IDs or HGNC Gene Symbols or
miRBase IDs). It is also possible to donwload annotations for protein
complexes. To do so, write "COMPLEX:" right before the genesymbols of
the genes integrating the complex. Check the vignette for examples.}

\item{filter_databases}{Load the annotations only from these databases.
See \code{\link{get_annotation_databases}} for possible values.}

\item{force_full_download}{Force the download of the entire annotations
dataset. This is disabled by default because the size of this data is
around 1GB. We recommend to retrieve the annotations for a set of proteins
or only from a few databases, depending on your interest.}

\item{...}{Additional arguments.}
}
\value{
A data.frame containing different gene/complex annotations
}
\description{
imports the annotations stored in Omnipath database from
\url{http://omnipathdb.org/annotations}
}
\examples{
annotations = import_Omnipath_annotations(select_genes=c("TP53","LMNA"),
     filter_databases=c("HPA_subcellular"))
}
\seealso{
\code{\link{get_annotation_databases}}
}
