% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_Omnipath_Interactions}
\alias{import_Omnipath_Interactions}
\title{Import Omnipath interaction database}
\usage{
import_Omnipath_Interactions(from_cache_file = NULL,
  filter_databases = .get_interaction_databases())
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are removed.}
}
\value{
data.frame
}
\description{
imports the database from http://omnipathdb.org/interactions, which contains
only interactions with references
}
\examples{
interactions = import_Omnipath_Interactions(filter_databases=c("SignaLink3","PhosphoSite", "Signor"))
}
