% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_Omnipath_PTMS}
\alias{import_Omnipath_PTMS}
\title{Import Omnipath PTMS}
\usage{
import_Omnipath_PTMS(from_cache_file = NULL,
  filter_databases = .get_ptms_databases())
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are removed.}
}
\value{
data.frame
}
\description{
imports the ptms database from http://omnipathdb.org/ptms
}
\examples{
ptms = import_Omnipath_PTMS(filter_databases=c("PhosphoSite", "Signor"))
}
