% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_PathwayExtra_Interactions}
\alias{import_PathwayExtra_Interactions}
\title{Imports from Omnipath webservice the interactions from 
Pathwayextra dataset}
\usage{
import_PathwayExtra_Interactions(from_cache_file = NULL,
  filter_databases = .get_interaction_databases(),
  select_organism = 9606)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are 
removed. See \code{\link{.get_interaction_databases}} for more information.}

\item{select_organism}{Interactions are available for human, mouse and rat. 
Choose one of those: 9606 human (default), 10116 rat or 10090 Mouse}
}
\value{
A dataframe containing activity flow interactions between proteins
without literature reference
}
\description{
Imports the dataset from: 
\url{http://omnipathdb.org/interactions?datasets=pathwayextra}, 
which contains activity flow interactions without literature reference
}
\examples{
interactions <- 
    import_PathwayExtra_Interactions(filter_databases=c("BioGRID","IntAct"),
     select_organism = 9606)
}
\seealso{
\code{\link{.get_interaction_databases}, 
  \link{import_AllInteractions}}
}
