% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{import_post_translational_interactions}
\alias{import_post_translational_interactions}
\title{All post-translational interactions from OmniPath}
\usage{
import_post_translational_interactions(
  resources = NULL,
  organism = "human",
  exclude = NULL,
  references_by_resource = TRUE,
  strict_evidences = FALSE,
  genesymbol_resource = NULL,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Character or integer: Name or NCBI Taxonomy ID of one or
organisms. The web service currently provides interactions for
human, mouse and rat. For other organisms, the data will be translated
by orthologous gene pairs from human. In this case, only one organism
can be provided.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{genesymbol_resource}{Character: either "uniprot" or "ensembl". The
former leaves intact the gene symbols returned by the web service,
originally set from UniProt. The latter updates the gene symbols from
Ensembl, which uses a slightly different gene symbol standard. In this
case a few records will be duplicated, where Ensembl provides ambiguous
translation.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing post-translational interactions
}
\description{
Imports interactions from all post-translational datasets of OmniPath.
The datasets are "omnipath", "kinaseextra", "pathwayextra" and
"ligrecextra".
}
\examples{
interactions <-
    import_post_translational_interactions(
        resources = c('BioGRID')
    )

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
