% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_tf_target_interactions}
\alias{import_tf_target_interactions}
\title{Imports interactions from the TF-target dataset of OmniPath}
\usage{
import_tf_target_interactions(
  resources = NULL,
  organism = 9606,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{...}{Optional additional arguments}
}
\value{
A dataframe containing TF-target interactions
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=tf_target},
which contains transcription factor-target protein coding gene
interactions. Note: this is not the only TF-target dataset in OmniPath,
`dorothea` is the other one and the `tf_mirna` dataset provides
TF-miRNA gene interactions.
}
\examples{
interactions <-
    import_tf_target_interactions(
        resources = c('DoRothEA', 'SIGNOR')
    )

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
