% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_TFregulons_Interactions}
\alias{import_TFregulons_Interactions}
\title{Imports from Omnipath webservice the interactions from 
Dorothea dataset}
\usage{
import_TFregulons_Interactions(
  from_cache_file = NULL,
  filter_databases = get_interaction_databases(),
  select_organism = 9606
)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are 
removed. See \code{\link{get_interaction_databases}} for more information.}

\item{select_organism}{Interactions are available for human, mouse and rat. 
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}
}
\value{
A dataframe containing TF-target interactions from DoRothEA
}
\description{
Imports the dataset from: 
\url{http://omnipathdb.org/interactions?datasets=tfregulons} 
which contains transcription factor (TF)-target interactions from DoRothEA
\url{https://github.com/saezlab/DoRothEA}
}
\examples{
interactions <- import_TFregulons_Interactions(filter_databases=c("DoRothEA_A",
    "ARACNe-GTEx_DoRothEA"), select_organism=9606)
}
\seealso{
\code{\link{get_interaction_databases}, 
  \link{import_AllInteractions}}
}
