% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{import_transcriptional_interactions}
\alias{import_transcriptional_interactions}
\title{Imports all TF-target interactions from OmniPath}
\usage{
import_transcriptional_interactions(
  resources = NULL,
  organism = "human",
  dorothea_levels = c("A", "B"),
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  genesymbol_resource = NULL,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Character or integer: Name or NCBI Taxonomy ID of one or
organisms. The web service currently provides interactions for
human, mouse and rat. For other organisms, the data will be translated
by orthologous gene pairs from human. In this case, only one organism
can be provided.}

\item{dorothea_levels}{Vector detailing the confidence levels of the
interactions to be downloaded. In dorothea, every TF-target interaction
has a confidence score ranging from A to E, being A the most reliable
interactions.
By default we take A and B level interactions (\code{c(A, B)}).
It is to note that E interactions are not available in OmnipathR.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{genesymbol_resource}{Character: either "uniprot" or "ensembl". The
former leaves intact the gene symbols returned by the web service,
originally set from UniProt. The latter updates the gene symbols from
Ensembl, which uses a slightly different gene symbol standard. In this
case a few records will be duplicated, where Ensembl provides ambiguous
translation.}

\item{...}{Optional additional arguments.}
}
\value{
A dataframe containing TF-target interactions.
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=tf_target,dorothea},
which contains transcription factor-target protein coding gene
interactions.
}
\examples{
grn <-
    import_transcriptional_interactions(
        resources = c('PAZAR', 'ORegAnno', 'DoRothEA')
    )
grn

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
