% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_datasets}
\alias{interaction_datasets}
\title{Datasets in the OmniPath Interactions database}
\usage{
interaction_datasets()
}
\value{
Character: labels of interaction datasets.
}
\description{
Datasets in the OmniPath Interactions database
}
\examples{
interaction_datasets()

}
