% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_resources}
\alias{interaction_resources}
\alias{get_interaction_resources}
\title{Interaction resources available in Omnipath}
\usage{
interaction_resources(dataset = NULL)
}
\arguments{
\item{dataset}{a dataset within the interactions query type. Currently
available datasets are `omnipath`, `kinaseextra`, `pathwayextra`,
`ligrecextra`, `collectri`, `dorothea`, `tf_target`, `tf_mirna`,
`mirnatarget`, `lncrna_mrna` and `small_molecule_protein`.}
}
\value{
Character: names of the interaction resources.
}
\description{
Names of the resources available in \url{https://omnipath.org/interactions}.
}
\examples{
interaction_resources()

}
\seealso{
\itemize{
    \item{\code{\link{resources}}}
    \item{\code{\link{omnipath}}}
    \item{\code{\link{pathwayextra}}}
    \item{\code{\link{kinaseextra}}}
    \item{\code{\link{ligrecextra}}}
    \item{\code{\link{post_translational}}}
    \item{\code{\link{dorothea}}}
    \item{\code{\link{collectri}}}
    \item{\code{\link{tf_target}}}
    \item{\code{\link{transcriptional}}}
    \item{\code{\link{mirna_target}}}
    \item{\code{\link{tf_mirna}}}
    \item{\code{\link{small_molecule}}}
    \item{\code{\link{all_interactions}}}
}
}
