% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_types}
\alias{interaction_types}
\title{Interaction types in the OmniPath Interactions database}
\usage{
interaction_types()
}
\value{
Character: labels of interaction types.
}
\description{
Interaction types in the OmniPath Interactions database
}
\examples{
interaction_types()

}
