% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{intercell_categories}
\alias{intercell_categories}
\title{Full list of intercell categories and resources}
\usage{
intercell_categories()
}
\value{
A data frame of categories and resources.
}
\description{
Full list of intercell categories and resources
}
\examples{
ic_cat <- intercell_categories()
ic_cat
# # A tibble: 1,125 x 3
#    category                parent                  database
#    <chr>                   <chr>                   <chr>
#  1 transmembrane           transmembrane           UniProt_location
#  2 transmembrane           transmembrane           UniProt_topology
#  3 transmembrane           transmembrane           UniProt_keyword
#  4 transmembrane           transmembrane_predicted Phobius
#  5 transmembrane_phobius   transmembrane_predicted Almen2009
# # . with 1,120 more rows

}
