% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercell.R
\name{intercell_generic_categories}
\alias{intercell_generic_categories}
\title{Retrieves a list of the generic categories in the intercell database
of OmniPath}
\usage{
intercell_generic_categories()
}
\value{
character vector with the different intercell main classes
}
\description{
Retrieves a list of the generic categories from
\url{https://omnipath.org/intercell}.
}
\examples{
get_intercell_generic_categories()

}
\seealso{
\itemize{
    \item{\code{\link{intercell}}}
    \item{\code{\link{intercell_categories}}}
    \item{\code{\link{intercell_summary}}}
}
}
