% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{is_ontology_id}
\alias{is_ontology_id}
\title{Looks like an ontology ID}
\usage{
is_ontology_id(terms)
}
\arguments{
\item{terms}{Character vector with strings to check.}
}
\value{
A logical vector with the same length as the input.
}
\description{
Tells if the input has the typical format of ontology IDs, i.e. a code
of capital letters, a colon, followed by a numeric code.
}
\examples{
is_ontology_id(c('GO:0000001', 'reproduction'))
# [1]  TRUE FALSE

}
