% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{is_swissprot}
\alias{is_swissprot}
\title{Check for SwissProt IDs}
\usage{
is_swissprot(uniprots, organism = 9606)
}
\arguments{
\item{uniprots}{Character vector of UniProt IDs.}

\item{organism}{Character or integer: name or identifier of the organism.}
}
\value{
Logical vector TRUE for SwissProt IDs and FALSE for any other
    element.
}
\description{
Check for SwissProt IDs
}
\examples{
is_swissprot(c("Q05BL1", "A0A654IBU3", "P00533"))
# [1] FALSE FALSE TRUE

}
