% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{is_trembl}
\alias{is_trembl}
\title{Check for TrEMBL IDs}
\usage{
is_trembl(uniprots, organism = 9606)
}
\arguments{
\item{uniprots}{Character vector of UniProt IDs.}

\item{organism}{Character or integer: name or identifier of the organism.}
}
\value{
Logical vector TRUE for TrEMBL IDs and FALSE for any other
    element.
}
\description{
Check for TrEMBL IDs
}
\examples{
is_trembl(c("Q05BL1", "A0A654IBU3", "P00533"))
# [1] TRUE TRUE FALSE

}
