% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_api_templates}
\alias{kegg_api_templates}
\title{List of templates in the KEGG REST API}
\usage{
kegg_api_templates()
}
\value{
A list of KEGG API templates.
}
\description{
List of templates in the KEGG REST API
}
\examples{
kegg_api_templates()

}
