% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_conv}
\alias{kegg_conv}
\title{Convert KEGG identifiers to/from outside identifiers}
\usage{
kegg_conv(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=kegg_query]{kegg_query}}
  \describe{
    \item{\code{operation}}{Character: one of the KEGG REST API operations.}
  }}
}
\value{
Data frame (tibble) of two columns with names "id_a" and "id_b".
}
\description{
See \url{https://www.kegg.jp/kegg/rest/keggapi.html#conv} for details.
}
\examples{
kegg_conv("compound", "pubchem")

}
