% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_databases}
\alias{kegg_databases}
\title{List of databases (endpoints) in the KEGG REST API}
\usage{
kegg_databases()
}
\value{
A character vector of KEGG databases.
}
\description{
List of databases (endpoints) in the KEGG REST API
}
\examples{
kegg_databases()

}
