% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{kegg_info}
\alias{kegg_info}
\title{Information about a KEGG Pathway}
\usage{
kegg_info(pathway_id)
}
\arguments{
\item{pathway_id}{Character: a KEGG Pathway identifier, e.g. "hsa04710".
For a complete list of IDs see \code{\link{kegg_pathway_list}}.}
}
\value{
List with the pathway information.
}
\description{
Information about a KEGG Pathway
}
\examples{
kegg_info('map00563')

}
\seealso{
\itemize{
    \item{\code{\link{kegg_pathway_list}}}
    \item{\code{\link{kegg_picture}}}
    \item{\code{\link{kegg_open}}}
}
}
