% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_list}
\alias{kegg_list}
\title{Obtain a list of KEGG entry identifiers and associated names}
\usage{
kegg_list(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=kegg_query]{kegg_query}}
  \describe{
    \item{\code{operation}}{Character: one of the KEGG REST API operations.}
  }}
}
\value{
Data frame (tibble) of two columns with names "id" and "name";
    except if the <database> argument is "organism", which results a
    three columns data frame ("genome", "kegg_name" and "latin_name").
}
\description{
See \url{https://www.kegg.jp/kegg/rest/keggapi.html#list} for details.
}
\examples{
kegg_list("pathway")

}
