% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_operations}
\alias{kegg_operations}
\title{List of operations in the KEGG REST API}
\usage{
kegg_operations()
}
\value{
A character vector of KEGG operations.
}
\description{
List of operations in the KEGG REST API
}
\examples{
kegg_operations()

}
