% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_organism_codes}
\alias{kegg_organism_codes}
\title{All 3 letter organism code from KEGG}
\usage{
kegg_organism_codes()
}
\value{
A character vector with all 3 letter codes.
}
\description{
All 3 letter organism code from KEGG
}
\examples{
kegg_organism_codes()

}
