% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_organisms}
\alias{kegg_organisms}
\title{List of organisms in KEGG}
\usage{
kegg_organisms()
}
\value{
A data frame (tibble) with organism data.
}
\description{
List of organisms in KEGG
}
\examples{
kegg_organisms()

}
