% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{kegg_pathway_annotations}
\alias{kegg_pathway_annotations}
\title{Protein pathway annotations}
\usage{
kegg_pathway_annotations(pathways = NULL)
}
\arguments{
\item{pathways}{A table of KEGG pathways as produced by \code{
\link{kegg_pathways_download}}.}
}
\value{
A data frame (tibble) with UniProt IDs and pathway names.
}
\description{
Downloads all KEGG pathways and creates a table of protein-pathway
annotations.
}
\examples{
\donttest{
kegg_pw_annot <- kegg_pathway_annotations()
}

}
\seealso{
\code{\link{kegg_pathways_download}}
}
