% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{kegg_pathway_list}
\alias{kegg_pathway_list}
\title{List of KEGG pathways}
\usage{
kegg_pathway_list()
}
\value{
Data frame of pathway names and identifiers.
}
\description{
Retrieves a list of available KEGG pathways.
}
\examples{
kegg_pws <- kegg_pathway_list()
kegg_pws
# # A tibble: 521 x 2
#    id       name
#    <chr>    <chr>
#  1 map01100 Metabolic pathways
#  2 map01110 Biosynthesis of secondary metabolites
#  3 map01120 Microbial metabolism in diverse environments
#  4 map01200 Carbon metabolism
#  5 map01210 2-Oxocarboxylic acid metabolism
#  6 map01212 Fatty acid metabolism
#  7 map01230 Biosynthesis of amino acids
# # . with 514 more rows

}
\seealso{
\itemize{
    \item{\code{\link{kegg_process}}}
    \item{\code{\link{kegg_pathway_download}}}
    \item{\code{\link{kegg_pathways_download}}}
}
}
