% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_request}
\alias{kegg_request}
\title{Perform a KEGG REST API request}
\usage{
kegg_request(operation, ...)
}
\arguments{
\item{operation}{Character: one of the KEGG REST API operations.}

\item{...}{Further parameters to \code{\link{kegg_request}}.}
}
\description{
Perform a KEGG REST API request
}
