% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinasephos.R
\name{kinasephos}
\alias{kinasephos}
\title{Kinase-PTM relationships from KinasePhos 3.0}
\usage{
kinasephos()
}
\value{
A data frame (tibble) of kinase-PTM interactions.
}
\description{
Predicted kinase-PTM interactions from https://awi.cuhk.edu.cn/KinasePhos/.
}
\examples{
kinasephos()

}
