% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo.R
\name{metabo_cosmos_organisms}
\alias{metabo_cosmos_organisms}
\title{Organisms with a COSMOS PKN available from omnipath-metabo}
\usage{
metabo_cosmos_organisms()
}
\value{
Integer vector of NCBI Taxonomy IDs.
}
\description{
Lists the NCBI Taxonomy IDs currently pre-built and cached by the
omnipath-metabo web service.
}
\examples{
\dontrun{
metabo_cosmos_organisms()
}

}
\seealso{
\itemize{
    \item{\code{\link{metabo_cosmos_pkn}}}
    \item{\code{\link{metabo_cosmos_categories}}}
}
}
