% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo_networks.R
\name{metabo_network_status}
\alias{metabo_network_status}
\title{Build status of a specialized network on omnipath-metabo}
\usage{
metabo_network_status(name)
}
\arguments{
\item{name}{Character: a network name, as listed by
\code{\link{metabo_networks}}.}
}
\value{
A named list with elements \code{name}, \code{present},
    \code{row_count} and \code{build_id}.
}
\description{
Build status of a specialized network on omnipath-metabo
}
\examples{
\dontrun{
metabo_network_status('metalinksdb')
}

}
\seealso{
\itemize{
    \item{\code{\link{metabo_networks}}}
    \item{\code{\link{metabo_network_interactions}}}
}
}
