% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatlas.R
\name{metabolic_atlas_list_gems}
\alias{metabolic_atlas_list_gems}
\title{List standard GEMs from Metabolic Atlas}
\usage{
metabolic_atlas_list_gems()
}
\value{
A data frame (tibble) with standard GEMs metadata and release information.
    Each model may have multiple rows, one for each release. The `latest` column
    indicates the most recent release for each model. The `repo_tree` column
    contains a character vector of file paths in the repository at the latest
    version, or NULL if the `gert` or `git2r` packages are not available.
}
\description{
Retrieves information about standard GEMs from Metabolic Atlas.
}
\examples{
\dontrun{
metabolic_atlas_list_gems()
}

}
