% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatlas.R
\name{metabolic_atlas_models}
\alias{metabolic_atlas_models}
\title{List of GEMs in Metabolic Atlas}
\usage{
metabolic_atlas_models()
}
\value{
A data frame (tibble) of GEMs.
}
\description{
Metabolic Atlas is a repository of genome scale models of metabolism (GEMs)
for various organisms, tissues and conditions. This function returns a list
of the available GEMs. Read more at \url{https://metabolicatlas.org}.
}
\examples{
metabolic_atlas_models()

}
