% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalinksdb.R
\name{metalinksdb_sqlite}
\alias{metalinksdb_sqlite}
\title{Open MetalinksDB as an SQLite3 connection}
\usage{
metalinksdb_sqlite()
}
\value{
An SQLite3 connection.
}
\description{
MetalinksDB is a database of metabolite-protein and small molecule
ligand-receptor interactions.
}
\examples{
con <- metalinksdb_sqlite()
con

}
