% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalinksdb.R
\name{metalinksdb_table}
\alias{metalinksdb_table}
\title{A table from MetalinksDB}
\usage{
metalinksdb_table(name)
}
\arguments{
\item{name}{Character. The name of the MetalinksDB table to fetch.}
}
\value{
A data frame (tibble) of one table from the MetalinksDB SQLite database.
}
\description{
A table from MetalinksDB
}
\examples{
metalinksdb_table('pathway')

}
\seealso{
\itemize{
    \item{\code{\link{metalinksdb_sqlite}}}
    \item{\code{\link{metalinksdb_tables}}}
}
}
