% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalinksdb.R
\name{metalinksdb_tables}
\alias{metalinksdb_tables}
\title{List tables in MetalinksDB}
\usage{
metalinksdb_tables()
}
\value{
Character vector of table names in the MetalinksDB SQLite database.
}
\description{
List tables in MetalinksDB
}
\examples{
metalinksdb_tables()

}
\seealso{
\itemize{
    \item{\code{\link{metalinksdb_sqlite}}}
}
}
