% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_gr_network_htridb}
\alias{nichenet_gr_network_htridb}
\title{NicheNet gene regulatory network from HTRIdb}
\usage{
nichenet_gr_network_htridb()
}
\description{
Builds a gene regulatory network using data from the HTRIdb database
and converts it to a format suitable for NicheNet.
}
\examples{
htri_gr_network <- nichenet_gr_network_htridb()

}
\seealso{
\code{\link{htridb_download}, \link{nichenet_gr_network}}
}
