% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_gr_network_pathwaycommons}
\alias{nichenet_gr_network_pathwaycommons}
\title{NicheNet gene regulatory network from PathwayCommons}
\usage{
nichenet_gr_network_pathwaycommons(
  interaction_types = "controls-expression-of",
  ...
)
}
\arguments{
\item{interaction_types}{Character vector with PathwayCommons interaction
types. Please refer to the default value and the PathwayCommons
webpage.}

\item{...}{Ignored.}
}
\description{
Builds gene regulation prior knowledge for NicheNet using PathwayCommons.
}
\examples{
pc_gr_network <- nichenet_gr_network_pathwaycommons()

}
\seealso{
\itemize{
    \item{\code{\link{nichenet_gr_network}}}
    \item{\code{\link{pathwaycommons_download}}}
}
}
