% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_gr_network_regnetwork}
\alias{nichenet_gr_network_regnetwork}
\title{NicheNet gene regulatory network from RegNetwork}
\usage{
nichenet_gr_network_regnetwork()
}
\value{
Data frame with gene regulatory interactions in NicheNet format.
}
\description{
Builds a gene regulatory network using data from the RegNetwork database
and converts it to a format suitable for NicheNet.
}
\examples{
regn_gr_network <- nichenet_gr_network_regnetwork()

}
\seealso{
\itemize{
    \item{\code{\link{regnetwork_download}}}
    \item{\code{\link{nichenet_gr_network}}}
}
}
