% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_gr_network_trrust}
\alias{nichenet_gr_network_trrust}
\title{NicheNet gene regulatory network from TRRUST}
\usage{
nichenet_gr_network_trrust()
}
\description{
Builds a gene regulatory network using data from the TRRUST database
and converts it to a format suitable for NicheNet.
}
\examples{
trrust_gr_network <- nichenet_gr_network_trrust()

}
\seealso{
\itemize{
    \item{\code{\link{trrust_download}}}
    \item{\code{\link{nichenet_gr_network}}}
}
}
