% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_lr_network_guide2pharma}
\alias{nichenet_lr_network_guide2pharma}
\title{Ligand-receptor network from Guide to Pharmacology}
\usage{
nichenet_lr_network_guide2pharma()
}
\value{
Data frame with ligand-receptor interactions in NicheNet format.
}
\description{
Downloads ligand-receptor interactions from the Guide to Pharmacology
database and converts it to a format suitable for NicheNet.
}
\examples{
g2p_lr_network <- nichenet_lr_network_guide2pharma()

}
\seealso{
\code{\link{nichenet_lr_network}, \link{guide2pharma_download}}
}
