% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_results_dir}
\alias{nichenet_results_dir}
\title{Path to the current NicheNet results directory}
\usage{
nichenet_results_dir()
}
\description{
Path to the directory to save intermediate and final outputs from NicheNet
methods.
}
\examples{
nichenet_results_dir()
# [1] "nichenet_results"

}
