% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_signaling_network_harmonizome}
\alias{nichenet_signaling_network_harmonizome}
\title{NicheNet signaling network from Harmonizome}
\usage{
nichenet_signaling_network_harmonizome(
  datasets = c("phosphositeplus", "kea", "depod"),
  ...
)
}
\arguments{
\item{datasets}{The datasets to use. For possible values please refer to
default value and the Harmonizome webpage.}

\item{...}{Ignored.}
}
\value{
A network data frame (tibble) with signaling interactions
    suitable for use with NicheNet.
}
\description{
Builds signaling network prior knowledge for NicheNet using Harmonizome
}
\examples{
# use only KEA and PhosphoSite:
hz_signaling_network <- nichenet_signaling_network_harmonizome(
    datasets = c('kea', 'phosphositeplus')
)

}
