% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_signaling_network_inbiomap}
\alias{nichenet_signaling_network_inbiomap}
\title{NicheNet signaling network from InWeb InBioMap}
\usage{
nichenet_signaling_network_inbiomap(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A network data frame (tibble) with signaling interactions
    suitable for use with NicheNet.
}
\description{
Builds signaling network prior knowledge for NicheNet from the InWeb
InBioMap database.
}
\examples{
\donttest{
ib_signaling_network <- nichenet_signaling_network_inbiomap()
}

}
\seealso{
\code{\link{nichenet_signaling_network}, \link{inbiomap_download}}
}
