% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_workarounds}
\alias{nichenet_workarounds}
\title{Workarounds using NicheNet without attaching the package}
\usage{
nichenet_workarounds()
}
\value{
Returns \code{NULL}.
}
\description{
NicheNet requires the availability of some lazy loaded external data
which are not available if the package is not loaded and attached. Also,
the \code{BBmisc::convertToShortString} used for error reporting in
\code{mlrMBO::evalTargetFun.OptState} is patched here to print longer
error messages. Maybe it's a better solution to attach \code{nichenetr}
before running the NicheNet pipeline. Alternatively you can try to call
this function in the beginning. Why we don't call this automatically is
just because we don't want to load datasets from another package without
the user knowing about it.
}
\examples{
\donttest{
nichenet_workarounds()
}

}
