% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oma.R
\name{oma_organisms}
\alias{oma_organisms}
\title{Organism identifiers from the Orthologous Matrix}
\usage{
oma_organisms()
}
\value{
A data frame with organism identifiers.
}
\description{
Organism identifiers from the Orthologous Matrix
}
\examples{
oma_organisms()

}
\seealso{
\code{\link{ensembl_organisms}}
}
