% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{omnipath}
\alias{omnipath}
\title{Literature curated signaling pathways}
\usage{
omnipath(
  resources = NULL,
  organism = 9606L,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  genesymbol_resource = NULL,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Character or integer: Name or NCBI Taxonomy ID of one or
organisms. The web service currently provides interactions for
human, mouse and rat. For other organisms, the data will be translated
by orthologous gene pairs from human. In this case, only one organism
can be provided.}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{genesymbol_resource}{Character: either "uniprot" or "ensembl". The
former leaves intact the gene symbols returned by the web service,
originally set from UniProt. The latter updates the gene symbols from
Ensembl, which uses a slightly different gene symbol standard. In this
case a few records will be duplicated, where Ensembl provides ambiguous
translation.}

\item{...}{optional additional arguments, passed to
\code{\link{import_omnipath_interactions}}.}
}
\value{
A dataframe of literature curated, post-translational signaling
    interactions.
}
\description{
Imports interactions from the `omnipath` dataset of Omnipath, a dataset
that inherits most of its design and contents from the original OmniPath
core from the 2016 publication. This dataset consists of about 40k
interactions.
}
\examples{
pathways <- omnipath()
pathways

}
\seealso{
\itemize{
    \item{\code{\link{import_omnipath_interactions}}}
    \item{\code{\link{import_post_translational_interactions}}}
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
