% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_autoclean}
\alias{omnipath_cache_autoclean}
\title{Keeps only the latest versions of complete downloads}
\usage{
omnipath_cache_autoclean()
}
\description{
Removes the old versions, the failed downloads and the files in the cache
directory which are missing from the database. For more flexible
operations use \code{\link{omnipath_cache_remove}} and
\code{\link{omnipath_cache_clean}}.
}
\examples{
\donttest{
omnipath_cache_autoclean()
}

}
