% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_clean}
\alias{omnipath_cache_clean}
\title{Removes the items from the cache directory which are unknown by the cache
database}
\usage{
omnipath_cache_clean()
}
\description{
Removes the items from the cache directory which are unknown by the cache
database
}
\examples{
omnipath_cache_clean()

}
