% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_clean_db}
\alias{omnipath_cache_clean_db}
\title{Removes the cache database entries without existing files}
\usage{
omnipath_cache_clean_db(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Returns `NULL`.
}
\description{
Removes the cache database entries without existing files
}
\examples{
omnipath_cache_clean_db()

}
