% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_key}
\alias{omnipath_cache_key}
\title{Generates a hash which identifies an element in the cache database}
\usage{
omnipath_cache_key(url, post = NULL, payload = NULL)
}
\arguments{
\item{url}{Character vector with URLs}

\item{post}{List with the HTTP POST parameters or a list of lists if
the url vector is longer than 1. NULL for queries without POST
parameters.}

\item{payload}{HTTP data payload. List with multiple items if the url
vector is longer than 1. NULL for queries without data.}
}
\value{
Character vector of cache record keys
}
\description{
Generates a hash which identifies an element in the cache database
}
\examples{
\donttest{
my_url <- 'https://bioconductor.org/'
omnipath_cache_key(my_url)
# [1] "41346a00fb20d2a9df03aa70cf4d50bf88ab154a"
}

}
