% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_latest_version}
\alias{omnipath_cache_latest_version}
\title{Finds the most recent version in a cache record}
\usage{
omnipath_cache_latest_version(record)
}
\arguments{
\item{record}{A cache record}
}
\value{
Character: the version ID with the most recent download finished
    time
}
\description{
Finds the most recent version in a cache record
}
