% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_move_in}
\alias{omnipath_cache_move_in}
\title{Moves an existing file into the cache}
\usage{
omnipath_cache_move_in(
  path,
  key = NULL,
  version = NULL,
  url = NULL,
  post = NULL,
  payload = NULL,
  keep_original = FALSE
)
}
\arguments{
\item{path}{Path to the source file}

\item{key}{Key of the cache item}

\item{version}{Version of the cache item. If does not exist a new version
item will be created}

\item{url}{URL of the downloaded resource}

\item{post}{HTTP POST parameters as a list}

\item{payload}{HTTP data payload}

\item{keep_original}{Whether to keep or remove the original file}
}
\description{
Either the key or the URL (with POST and payload) must be provided.
}
\examples{
\donttest{
omnipath_cache_move_in('some/file.zip', url = 'the_download_address')
}

}
\seealso{
\code{\link{omnipath_cache_save}}
}
