% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_search}
\alias{omnipath_cache_search}
\title{Searches for cache items}
\usage{
omnipath_cache_search(pattern, ...)
}
\arguments{
\item{pattern}{String or regular expression.}

\item{...}{Passed to \code{\link{grep}}}
}
\value{
List of cache records matching the pattern.
}
\description{
Searches the cache records by matching the URL against a string or regexp.
}
\examples{
# find all cache records from the BioPlex database
bioplex_cache_records <- omnipath_cache_search(
    'bioplex',
    ignore.case = TRUE
)

}
