% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_set_ext}
\alias{omnipath_cache_set_ext}
\title{Sets the file extension for a cache record}
\usage{
omnipath_cache_set_ext(key, ext)
}
\arguments{
\item{key}{Character: key for a cache item, alternatively a version entry.}

\item{ext}{Character: the file extension, e.g. "zip".}
}
\description{
Sets the file extension for a cache record
}
\examples{
\donttest{
my_url <- 'https://bioconductor.org/'
key <- omnipath_cache_key(url = my_url)
omnipath_cache_set_ext(key = key, ext = 'html')
}

}
