% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_cache_wipe}
\alias{omnipath_cache_wipe}
\title{Permanently removes all the cache contents}
\usage{
omnipath_cache_wipe(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Returns `NULL`.
}
\description{
After this operation the cache directory will be completely empty,
except an empty cache database file.
}
\examples{
\donttest{
omnipath_cache_wipe()
# the cache is completely empty:
print(omnipath.env$cache)
# list()
list.files(omnipath_get_cachedir())
# [1] "cache.json"
}


}
\seealso{
\code{\link{omnipath_cache_remove}}
}
