% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{omnipath_config_path}
\alias{omnipath_config_path}
\alias{config_path}
\title{Current config file path of OmnipathR}
\usage{
omnipath_config_path(user = FALSE)

config_path(user = FALSE, pkg = "OmnipathR")
}
\arguments{
\item{user}{Logical: prioritize the user level config even if a config in
the current working directory is available.}
}
\value{
Character: path to the config file.
}
\description{
Current config file path of OmnipathR

Current config file path for a certain package
}
\examples{
omnipath_config_path()

}
