% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmanager.R
\name{omnipath_db_available}
\alias{omnipath_db_available}
\title{Built in database definitions}
\usage{
omnipath_db_available()
}
\value{
A data frame with the build in databaase definitions.
}
\description{
Databases are resources which might be costly to load but can be used many
times by functions which usually automatically load and retrieve them from
the database manager. Each database has a lifetime and will be unloaded
automatically upon expiry.
}
