% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2graph.R
\name{omnipath_graph}
\alias{omnipath_graph}
\title{Build Omnipath interaction graph}
\usage{
omnipath_graph(interactions)
}
\arguments{
\item{interactions}{data.frame created by import_Omnipath_Interactions}
}
\value{
igraph object
}
\description{
transforms the interactions data.frame to igraph object
}
\examples{
interactions = import_Omnipath_Interactions(filter_databases=c("SignaLink3","PhosphoSite", "Signor"))
OPI_g = omnipath_graph(interactions = interactions )
}
