% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{omnipath_log}
\alias{omnipath_log}
\alias{read_log}
\title{Browse the current OmnipathR log file}
\usage{
omnipath_log()

read_log(pkg = "OmnipathR")
}
\arguments{
\item{pkg}{Character: name of a package.}
}
\value{
Returns `NULL`.
}
\description{
Browse the current OmnipathR log file

Browse the latest log from a package
}
\examples{
\dontrun{
omnipath_log()
# then you can browse the log file, and exit with `q`
}

}
\seealso{
\code{\link{omnipath_logfile}}
}
