% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{omnipath_logfile}
\alias{omnipath_logfile}
\title{Returns the path to the current OmnipathR log file}
\usage{
omnipath_logfile()
}
\value{
Character: path to the current logfile.
}
\description{
Returns the path to the current OmnipathR log file
}
\examples{
omnipath_logfile()
# [1] "/home/denes/omnipathr/omnipathr-log/omnipathr-20210309-1642.log"

}
\seealso{
\code{\link{omnipath_log}}
}
