% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{omnipath_set_console_loglevel}
\alias{omnipath_set_console_loglevel}
\title{Sets the log level for the console}
\usage{
omnipath_set_console_loglevel(level)
}
\arguments{
\item{level}{Character or class `loglevel`. The desired log level.}
}
\value{
Returns `NULL`.
}
\description{
Use this method to change during a session which messages you want to be
printed on the console. Before loading the package, you can set it also by
the config file, with the omnipath.console_loglevel key.
}
\examples{
omnipath_set_console_loglevel('warn')
# or:
omnipath_set_console_loglevel(logger::WARN)

}
\seealso{
\code{\link{omnipath_set_logfile_loglevel}}
}
