% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{omnipath_set_logfile_loglevel}
\alias{omnipath_set_logfile_loglevel}
\title{Sets the log level for the logfile}
\usage{
omnipath_set_logfile_loglevel(level)
}
\arguments{
\item{level}{Character or class `loglevel`. The desired log level.}
}
\description{
Use this method to change during a session which messages you want to be
written into the logfile. Before loading the package, you can set it also
by the config file, with the omnipath.loglevel key.
}
\examples{
omnipath_set_logfile_loglevel('info')
# or:
omnipath_set_logfile_loglevel(logger::INFO)

}
\seealso{
\code{\link{omnipath_set_console_loglevel}}
}
