% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmanager.R
\name{omnipath_show_db}
\alias{omnipath_show_db}
\title{Built in database definitions}
\usage{
omnipath_show_db()
}
\value{
A data frame with the build in databaase definitions.
}
\description{
Databases are resources which might be costly to load but can be used many
times by functions which usually automatically load and retrieve them from
the database manager. Each database has a lifetime and will be unloaded
automatically upon expiry.
}
\examples{
database_definitions <- omnipath_show_db()
database_definitions
# # A tibble: 14 x 10
#    name    last_used lifetime package loader loader_param latest_param
#    <chr>   <lgl>        <dbl> <chr>   <chr>  <list>       <lgl>
#  1 Gene O. NA             300 Omnipa. go_on. <named list. NA
#  2 Gene O. NA             300 Omnipa. go_on. <named list. NA
#  3 Gene O. NA             300 Omnipa. go_on. <named list. NA
#  4 Gene O. NA             300 Omnipa. go_on. <named list. NA
#  5 Gene O. NA             300 Omnipa. go_on. <named list. NA
# # . with 3 more variables: loaded <lgl>, db <lgl>, key <chr>

}
