% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{omnipath_unlock_cache_db}
\alias{omnipath_unlock_cache_db}
\title{Removes the lock file from the cache directory}
\usage{
omnipath_unlock_cache_db()
}
\value{
Logical: returns TRUE if the cache was locked and now is
    unlocked; FALSE if it was not locked.
}
\description{
A lock file in the cache directory avoids simulatneous write and read.
It's supposed to be removed after each read and write operation. This
might not happen if the process crashes during such an operation. In
this case you can manually call this function.
}
