% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthology.R
\name{orthology_translate}
\alias{orthology_translate}
\title{Translate identifiers between organisms by orthologous gene pairs}
\usage{
orthology_translate(data, ..., target = 10090, source = 9606)
}
\arguments{
\item{data}{Data frame or character vector.}

\item{...}{Column specification: from zero to up to three arguments, with
or without names. NSE is supported. Arguments beyond the third one
will be ignored.
\itemize{
    \item{
        The name of the arguments should be column names, the
        values identifier types, either as character or as symbols.
    }
    \item{
        Arguments without names assumed to be both column names and
        identifier types, e.g. a column called "uniprot" containing
        UniProt IDs.
    }
    \item{
        The first column spefication describes the source column,
        with identifiers of the source organism. This column must
        exist in the data and this will be the input of the homology
        translation. This column will be removed from the returned
        data frame.
    }
    \item{
        In case of "uniprot", the source column name can be anything,
        if it contains only UniProt IDs it will be handled accordingly.
    }
    \item{
        In case of "genesymbol", is enough if the source column name
        contains the word "genesymbol", e.g. "ligand_genesymbol".
    }
    \item{
        The second column spefication describes the target column,
        with its name and identifier type. If not provided, both
        the column name and type will be the same as the source
    }
    \item{
        Optionally a third column can be specified with another
        identifier type. This is convenient if you want, for example
        also Gene Symbols along with UniProt IDs.
    }
    \item{
        If no specification provided, the input assumed to have
        a column named either "uniprot" or "genesymbol", or be
        a character vector of UniProt IDs or Gene Symbols.
    }
}}

\item{target}{Character or integer: name or identifier of the target
organism (the one we translate to). The default target organism is
mouse.}

\item{source}{Character or integer: name of identifier of the source
organism (the one the IDs in the input data belong to). The default
source organism is human.}
}
\description{
Translates identifiers between organisms using orthology data from NCBI
HomoloGene.
}
