% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthology.R
\name{orthology_translate_column}
\alias{orthology_translate_column}
\title{Translate a column of identifiers by orthologous gene pairs}
\usage{
orthology_translate_column(
  data,
  column,
  id_type,
  target_organism,
  source_organism = 9606L,
  replace = FALSE,
  one_to_many = NULL,
  keep_untranslated = FALSE,
  translate_complexes = FALSE,
  uniprot_by_id_type = "entrez"
)
}
\arguments{
\item{data}{A data frame with the column to be translated.}

\item{column}{Name of a character column with identifiers of the source
organism of type `id_type`.}

\item{target_organism}{Name or NCBI Taxonomy ID of the target organism.}

\item{source_organism}{Name or NCBI Taxonomy ID of the source organism.}

\item{replace}{Logical or character: replace the column with the translated
identifiers, or create a new column. If it is character, the}

\item{one_to_many}{Integer: maximum number of orthologous pairs for one
gene of the source organism.}

\item{keep_untranslated}{Logical: keep records without orthologous pairs.
If `replace` is TRUE, this option is ignoredx, and untranslated records
will be dropped.}

\item{translate_complexes}{Logical: translate the complexes by translating
their components.}

\item{uniprot_by_id_type}{Character: translate NCBI HomoloGene to UniProt
by this ID type. One of "genesymbol", "entrez", "refseq" or "gi".}
}
\value{
The data frame with identifiers translated to other organism.
}
\description{
Translate a column of identifiers by orthologous gene pairs
}
