% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwaycommons.R
\name{pathwaycommons_download}
\alias{pathwaycommons_download}
\title{Interactions from PathwayCommons}
\usage{
pathwaycommons_download()
}
\description{
PathwayCommons (http://www.pathwaycommons.org/) provides molecular
interactions from a number of databases, in either BioPAX or SIF (simple
interaction format). This function retrieves all interactions in SIF
format. The data is limited to the interacting pair and the type of the
interaction.
}
