% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preppi.R
\name{preppi_download}
\alias{preppi_download}
\title{Interactions from PrePPI}
\usage{
preppi_download(...)
}
\arguments{
\item{...}{Minimum values for the scores. The available scores are:
str, protpep, str_max, red, ort, phy, coexp, go, total, exp and final.}
}
\value{
A data frame (tibble) of interactions with scores, databases
    and literature references.
}
\description{
Retrieves predicted protein-protein interactions from the PrePPI
database (\url{http://honig.c2b2.columbia.edu/preppi}). The interactions
in this table are supposed to be correct with a > 0.5 probability.
}
\examples{
preppi <- preppi_download()

}
