% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{print_interactions}
\alias{print_interactions}
\title{print interactions}
\usage{
print_interactions(interDF, writeRefs = F)
}
\arguments{
\item{interDF}{data.frame with the interacgtions}

\item{writeRefs}{[FALSE] writes also the PubMed IDs}
}
\description{
prints the interactions in a nice format
}
\examples{
ptms = import_Omnipath_PTMS()
print_interactions(head(ptms))
print_interactions(tail(ptms),writeRefs=T)
print_interactions(dplyr::filter(ptms,enzyme_genesymbol=="MAP2K1",substrate_genesymbol=="MAPK3"))
}
