% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{printPath_es}
\alias{printPath_es}
\title{print network paths given by edge sequence}
\usage{
printPath_es(edgeSeq, G)
}
\arguments{
\item{edgeSeq}{edge sequence}

\item{G}{igraph object (from ptms or any interaction dataset)}
}
\value{
Interactions displayed in a nice format
}
\description{
prints the interactions in the path in a nice format
}
\examples{
interactions = import_omnipath_interactions(resources=c("SignaLink3"))
OPI_g = interaction_graph(interactions = interactions )
printPath_es(shortest_paths(OPI_g,from = "TYRO3",to = "STAT3",
    output = 'epath')$epath[[1]],OPI_g)
}
\seealso{
\code{\link{printPath_vs}}
}
