% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{printPath_vs}
\alias{printPath_vs}
\title{print networks paths given by node sequence}
\usage{
printPath_vs(nodeSeq, G)
}
\arguments{
\item{nodeSeq}{node sequence}

\item{G}{igraph object (from ptms or interactions)}
}
\description{
prints the interactions in the path in a nice format
}
\examples{
interactions = import_Omnipath_Interactions(filter_databases=c("SignaLink3"))
OPI_g = interaction_graph(interactions = interactions )
printPath_vs(shortest_paths(OPI_g,from = "TYRO3",to = "STAT3")$vpath,OPI_g)

ptms = import_Omnipath_PTMS(filter_databases=c("PhosphoSite", "Signor"))
ptms_g = ptms_graph(ptms)
printPath_vs(all_shortest_paths(ptms_g,from = "SRC",to = "STAT1")$res,ptms_g)
}
\seealso{
\code{\link{printPath_es}}
}
