% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ptms_graph}
\alias{ptms_graph}
\title{Post-translational modifications (PTMs) graph}
\usage{
ptms_graph(ptms)
}
\arguments{
\item{ptms}{data.frame created by \code{\link{import_omnipath_enzsub}}}
}
\value{
An igraph object
}
\description{
transforms the ptms interactions data.frame to igraph object
}
\examples{
ptms = import_omnipath_enzsub(resources=c('PhosphoSite', 'SIGNOR'))
ptms_g = ptms_graph(ptms = ptms )
}
\seealso{
\code{\link{import_omnipath_enzsub}}
}
