% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pubmed_open}
\alias{pubmed_open}
\title{Open one or more PubMed articles}
\usage{
pubmed_open(pmids, browser = NULL, sep = ";", max_pages = 25L)
}
\arguments{
\item{pmids}{Character or numberic vector of one or more PubMed IDs.}

\item{browser}{Character: name of the web browser executable. If `NULL`,
the default web browser will be used.}

\item{sep}{Character: split the PubMed IDs by this separator.}

\item{max_pages}{Numeric: largest number of pages to open. This is to
prevent opening hundreds or thousands of pages at once.}
}
\value{
Returns `NULL`.
}
\description{
Open one or more PubMed articles
}
\examples{
interactions <- import_omnipath_interactions()
pubmed_open(interactions$references[1])

}
