% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{query_info}
\alias{query_info}
\title{OmniPath query parameters}
\usage{
query_info(query_type)
}
\arguments{
\item{query_type}{Character: interactions, annotations, complexes, enz_sub
or intercell.}
}
\description{
All parameter names and their possible values for a query type. Note:
parameters with `NULL` values have too many possible values to list
them.
}
\examples{
query_info()

}
