% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramilowski.R
\name{ramilowski_download}
\alias{ramilowski_download}
\title{Downloads ligand-receptor interactions from Ramilowski et al. 2015}
\usage{
ramilowski_download()
}
\description{
Curated ligand-receptor pairs from Supplementary Table 2 of the article
"A draft network of ligand-receptor mediated multicellular signaling in
human" (https://www.nature.com/articles/ncomms8866).
}
\examples{
rami_interactions <- ramilowski_download()

}
