% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp.R
\name{ramp_id_mapping_table}
\alias{ramp_id_mapping_table}
\title{Pairwise ID translation table from RaMP database}
\usage{
ramp_id_mapping_table(from, to, version = "2.5.4")
}
\arguments{
\item{from}{Character or Symbol. Name of an identifier type.}

\item{to}{Character or Symbol. Name of an identifier type.}

\item{version}{Character. The version of RaMP to download.}
}
\value{
Dataframe of pairs of identifiers.
}
\description{
Pairwise ID translation table from RaMP database
}
\examples{
ramp_id_mapping_table('hmdb', 'kegg')

}
\seealso{
\itemize{
    \item{\code{\link{ramp_sqlite}}}
    \item{\code{\link{ramp_tables}}}
    \item{\code{\link{ramp_table}}}
}
}
