% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{ramp_id_type}
\alias{ramp_id_type}
\title{RaMP identifier type label}
\usage{
ramp_id_type(label)
}
\arguments{
\item{label}{Character: an ID type label, as shown in the table returned 
by \code{\link{id_types}}}
}
\value{
Character: the RaMP specific ID type label, or the input
    unchanged if it could not be translated (still might be a valid
    identifier name). These labels should be valid value
    names, as used in RaMP SQL database.
}
\description{
RaMP identifier type label
}
\examples{
ramp_id_type("rhea")
# [1] "rhea-comp"

}
\seealso{
\itemize{
    \item{\code{\link{chalmers_gem_id_type}}}
    \item{\code{\link{uniprot_id_type}}}
    \item{\code{\link{ensembl_id_type}}}
    \item{\code{\link{uploadlists_id_type}}}
}
}
