% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp.R
\name{ramp_sqlite}
\alias{ramp_sqlite}
\title{Download and open RaMP database SQLite}
\usage{
ramp_sqlite(version = RAMP_LATEST_VERSION)
}
\arguments{
\item{version}{Character. The version of RaMP to download.}
}
\value{
SQLite connection.
}
\description{
Download and open RaMP database SQLite
}
\examples{
sqlite_con <- ramp_sqlite()

}
\seealso{
\itemize{
    \item{\code{\link{ramp_tables}}}
}
}
