% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp.R
\name{ramp_tables}
\alias{ramp_tables}
\title{List tables in RaMP database}
\usage{
ramp_tables(version = RAMP_LATEST_VERSION)
}
\arguments{
\item{version}{Character. The version of RaMP to download.}
}
\value{
Character vector of table names in the RaMP SQLite database.
}
\description{
List tables in RaMP database
}
\examples{
ramp_tables()

}
\seealso{
\itemize{
    \item{\code{\link{ramp_sqlite}}}
}
}
