% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{reactome_pathway_relations}
\alias{reactome_pathway_relations}
\title{Reactome pathway relations}
\usage{
reactome_pathway_relations(organism = NULL)
}
\arguments{
\item{organism}{Character or integer: organism name, common name or
NCBI Taxonomy ID. If \code{NULL}, returns all organisms.}
}
\value{
A tibble with columns \code{Parent}, \code{Child}.
}
\description{
Retrieves parent-child relationships between Reactome pathways and
optionally filters by organism.
}
\examples{
\dontrun{
relations <- reactome_pathway_relations(9606)
head(relations)
}
}
