% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon3d.R
\name{recon3d_metabolites}
\alias{recon3d_metabolites}
\alias{recon3d_reactions}
\alias{recon3d_genes}
\alias{recon3d_compartments}
\title{Metabolites from Recon-3D}
\usage{
recon3d_metabolites(extra_hmdb = TRUE)

recon3d_reactions()

recon3d_genes()

recon3d_compartments()
}
\arguments{
\item{extra_hmdb}{Logical: add extra HMDB IDs from Virtual Metabolic Human.}
}
\value{
Data frame: tibble of metabolites.

Data frame: tibble of reactions.

Data frame: tibble of genes.

Data frame: tibble of compartments.
}
\description{
Metabolites from Recon-3D

Reactions from Recon-3D

Genes from Recon-3D

Compartments from Recon-3D
}
\examples{
recon3d_metabolites()

recon3d_reactions()

recon3d_genes()

recon3d_compartments()

}
