% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon3d.R
\name{recon3d_raw_matlab}
\alias{recon3d_raw_matlab}
\title{Recon3D model from BiGG}
\usage{
recon3d_raw_matlab()
}
\value{
List: the Recon3D model as read from the Matlab file.
}
\description{
Returns the content extracted from Matlab file.
}
