% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon3d.R
\name{recon3d_raw_vmh}
\alias{recon3d_raw_vmh}
\title{Recon-3D model from Virtual Metabolic Human}
\usage{
recon3d_raw_vmh()
}
\value{
List: the Recon3D model as parsed from the VMH SBML file.
}
\description{
Recon-3D model from Virtual Metabolic Human
}
\examples{
recon3d_raw_vmh()

}
