% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnetwork.R
\name{regnetwork_directions}
\alias{regnetwork_directions}
\title{Transcription factor effects from RegNetwork}
\usage{
regnetwork_directions(organism = "human")
}
\arguments{
\item{organism}{Character: either human or mouse.}
}
\description{
Transcription factor effects from RegNetwork
}
\examples{
regn_dir <- regnetwork_directions()

}
