% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resource_info}
\alias{resource_info}
\title{OmniPath resource information}
\usage{
resource_info()
}
\description{
The `resources` query type provides resource metadata in JSON format.
Here we retrieve this JSON and return it as a nested list structure.
}
\examples{
resource_info()

}
