% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources_in}
\alias{resources_in}
\title{Collect resource names from a data frame}
\usage{
resources_in(data)
}
\arguments{
\item{data}{A data frame from an OmniPath query.}
}
\value{
Character: resource names occuring in the data frame.
}
\description{
Collect resource names from a data frame
}
\examples{
pathways <- omnipath_interactions()
resources_in(pathways)

}
