% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_tables}
\alias{static_tables}
\title{List the static tables available from OmniPath}
\usage{
static_tables()
}
\value{
A data frame listing the available tables.
}
\description{
A few resources and datasets are available also as plain TSV files and
can be accessed without TLS. The purpose of these tables is to make the
most often used OmniPath data available on computers with configuration
issues. These tables are not the recommended way to access OmniPath
data, and a warning is issued each time they are accessed.
}
\examples{
static_tables()

}
\seealso{
\code{\link{static_table}}
}
