% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_actions}
\alias{stitch_actions}
\title{Retrieve the STITCH actions dataset}
\usage{
stitch_actions(organism = "human", prefixes = FALSE)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier.
Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus),
10116 (Rattus norvegicu), 7955 (Danio rerio), 7227 (Drosophila
melanogaster) and 6239 (Caenorhabditis elegans).}
}
\value{
Data frame of STITCH actions.
}
\description{
Retrieve the STITCH actions dataset
}
\examples{
sta <- stitch_actions(organism = 'mouse')

}
\seealso{
\itemize{
    \item{\code{\link{stitch_actions}}}
    \item{\code{\link{stitch_links}}}
    \item{\code{\link{stitch_network}}}
}
}
