% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_gem}
\alias{stitch_gem}
\title{Chemical-protein interactions from STITCH}
\usage{
stitch_gem(
  organism = "human",
  min_score = 700L,
  protein_ids = c("uniprot", "genesymbol"),
  metabolite_ids = c("hmdb", "kegg"),
  cosmos = FALSE
)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier.
Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus),
10116 (Rattus norvegicu), 7955 (Danio rerio), 7227 (Drosophila
melanogaster) and 6239 (Caenorhabditis elegans).}

\item{min_score}{Confidence cutoff used for STITCH connections
(700 by default).}

\item{protein_ids}{Character: translate the protein identifiers to these ID
types. Each ID type results two extra columns in the output, for the "a"
and "b" sides of the interaction, respectively. The default ID type for
proteins is Esembl Protein ID, and by default UniProt IDs and Gene
Symbols are included.}

\item{metabolite_ids}{Character: translate the protein identifiers to these ID
types. Each ID type results two extra columns in the output, for the "a"
and "b" sides of the interaction, respectively. The default ID type for
metabolites is PubChem CID, and HMDB IDs and KEGG IDs are included.}

\item{cosmos}{Logical: use COSMOS format?}
}
\value{
A data frame of STITCH chemical-protein and protein-chemical
interactions with their effect signs, and optionally with identifiers
translated.
}
\description{
Chemical-protein interactions from STITCH
}
\examples{
stg <- stitch_gem(protein_ids = 'genesymbol', metabolite_ids = 'hmdb')

}
\seealso{
\itemize{
    \item{\code{\link{stitch_actions}}}
    \item{\code{\link{stitch_links}}}
    \item{\code{\link{stitch_remove_prefixes}}}
}
}
