% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_links}
\alias{stitch_links}
\title{Alberto Valdeolivas
Retrieve the STITCH links dataset}
\usage{
stitch_links(organism = "human", prefixes = FALSE)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier.
Supported taxons are 9606 (Homo sapiens), 10090 (Mus musculus),
10116 (Rattus norvegicu), 7955 (Danio rerio), 7227 (Drosophila
melanogaster) and 6239 (Caenorhabditis elegans).}
}
\value{
Data frame: organism specific STITCH links dataset.
}
\description{
Alberto Valdeolivas
Retrieve the STITCH links dataset
}
\examples{
stl <- stitch_links()

}
\seealso{
\itemize{
    \item{\code{\link{stitch_actions}}}
    \item{\code{\link{stitch_links}}}
    \item{\code{\link{stitch_gem}}}
}
}
