% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{trembls_only}
\alias{trembls_only}
\title{Retain only TrEMBL IDs}
\usage{
trembls_only(uniprots, organism = 9606)
}
\arguments{
\item{uniprots}{Character vector of UniProt IDs.}

\item{organism}{Character or integer: name or identifier of the organism.}
}
\value{
Character vector with only TrEMBL IDs.
}
\description{
Retain only TrEMBL IDs
}
\examples{
trembls_only(c("Q05BL1", "A0A654IBU3", "P00533"))
# [1] "Q05BL1" "A0A654IBU3"

}
