% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_id_mapping_table}
\alias{uniprot_id_mapping_table}
\title{Retrieves an identifier translation table from the UniProt uploadlists
service}
\usage{
uniprot_id_mapping_table(identifiers, from, to)
}
\arguments{
\item{identifiers}{Character vector of identifiers}

\item{from}{Type of the identifiers provided. See Details for possible
values.}

\item{to}{Identifier type to be retrieved from UniProt. See Details for
possible values.}
}
\value{
A data frame (tibble) with columns `From` and `To`, the
    identifiers provided and the corresponding target IDs, respectively.
}
\description{
Retrieves an identifier translation table from the UniProt uploadlists
service
}
\details{
This function uses the uploadlists service of UniProt to obtain identifier
translation tables. The possible values for `from` and `to` are the
identifier type abbreviations used in the UniProt API, please refer to
the table here: \url{https://www.uniprot.org/help/api_idmapping}
}
\examples{
uniprot_genesymbol <- uniprot_id_mapping_table(
    c('P00533', 'P23771'), 'ID', 'GENENAME'
)
uniprot_genesymbol
# # A tibble: 2 x 2
#   From   To
#   <chr>  <chr>
# 1 P00533 EGFR
# 2 P23771 GATA3
}
