% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{uniprot_id_type}
\alias{uniprot_id_type}
\title{UniProt identifier type label}
\usage{
uniprot_id_type(label)
}
\arguments{
\item{label}{Character: an ID type label, as shown in the table at
\code{\link{translate_ids}}}
}
\value{
Character: the UniProt specific ID type label, or the input
    unchanged if it could not be translated (still might be a valid
    identifier name). This is the label that one can use in UniProt
    REST queries.
}
\description{
UniProt identifier type label
}
\examples{
ensembl_id_type("entrez")
# [1] "database(GeneID)"

}
\seealso{
\itemize{
    \item{\code{\link{ensembl_id_type}}}
    \item{\code{\link{uploadlists_id_type}}}
}
}
