% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{uniprot_idmapping_id_types}
\alias{uniprot_idmapping_id_types}
\title{ID types available in the UniProt ID Mapping service}
\usage{
uniprot_idmapping_id_types()
}
\description{
ID types available in the UniProt ID Mapping service
}
\examples{
uniprot_idmapping_id_types()

}
