% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_organisms}
\alias{uniprot_organisms}
\title{UniProt taxonomy data}
\usage{
uniprot_organisms()
}
\value{
A tibble with columns: \code{code}, \code{kingdom}, \code{ncbi_tax_id}, \code{latin_name}, \code{common_name}, \code{synonym}.
}
\description{
UniProt taxonomy data
}
\examples{
uniprot_organisms()

}
