% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidences.R
\name{unnest_evidences}
\alias{unnest_evidences}
\title{Separate evidences by direction and effect sign}
\usage{
unnest_evidences(data, longer = FALSE)
}
\arguments{
\item{data}{An interaction data frame with "evidences" column.}

\item{longer}{Logical: If TRUE, the "evidences" column is split into rows.}
}
\value{
The data frame with new columns or new rows by direction and sign.
}
\description{
Separate evidences by direction and effect sign
}
\examples{
\dontrun{
omnipath <- import_omnipath_interactions(fields = 'evidences')
omnipath <- unnest_evidences(omnipath)
colnames(omnipath)
}

}
\seealso{
\itemize{
    \item{\code{\link{only_from}}}
    \item{\code{\link{filter_evidences}}}
    \item{\code{\link{from_evidences}}}
}
}
