% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{uploadlists_id_type}
\alias{uploadlists_id_type}
\title{UniProt Uploadlists identifier type label}
\usage{
uploadlists_id_type(label, side = "from")
}
\arguments{
\item{label}{Character: an ID type label, as shown in the table at
\code{\link{translate_ids}}}

\item{side}{Character: either "from" or "to": direction of the mapping.}
}
\value{
Character: the UniProt Uploadlists specific ID type label, or
    the input unchanged if it could not be translated (still might be
    a valid identifier name). This is the label that one can use in
    UniProt Uploadlists (ID Mapping) queries.
}
\description{
UniProt Uploadlists identifier type label
}
\examples{
ensembl_id_type("entrez")
# [1] "GeneID"

}
\seealso{
\itemize{
    \item{\code{\link{ensembl_id_type}}}
    \item{\code{\link{uniprot_id_type}}}
    \item{\code{\link{hmdb_id_type}}}
    \item{\code{\link{chalmers_gem_id_type}}}
}
}
