% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinayagam.R
\name{vinayagam_download}
\alias{vinayagam_download}
\title{Protein-protein interactions from Vinayagam 2011}
\usage{
vinayagam_download()
}
\description{
Retrieves the Supplementary Table S6 from Vinayagam et al. 2011.
Find out more at https://doi.org/10.1126/scisignal.2001699
}
\examples{
vinayagam_interactions <- vinayagam_download()
vinayagam_interactions
# # A tibble: 34,814 x 5
#    `Input-node Gen. `Input-node Gen. `Output-node Ge. `Output-node Ge.
#    <chr>                       <dbl> <chr>                       <dbl>
#  1 C1orf103                    55791 MNAT1                        4331
#  2 MAST2                       23139 DYNLL1                       8655
#  3 RAB22A                      57403 APPL2                       55198
#  4 TRAP1                       10131 EXT2                         2132
#  5 STAT2                        6773 COPS4                       51138
# # . with 34,804 more rows, and 1 more variable:
# # `Edge direction score` <dbl>

}
