% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{with_extra_attr}
\alias{with_extra_attr}
\title{Interaction records having an extra attribute}
\usage{
with_extra_attr(data, attr)
}
\arguments{
\item{data}{An interaction data frame.}

\item{attr}{The name of an extra attribute; NSE is supported.}
}
\value{
The data frame filtered to the records having the extra attribute.
}
\description{
Interaction records having an extra attribute
}
\examples{
i <- import_omnipath_interactions(fields = 'extra_attrs')
with_extra_attr(i, Macrophage_type)

}
