% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{with_extra_attrs}
\alias{with_extra_attrs}
\title{Interaction records having certain extra attributes}
\usage{
with_extra_attrs(data, ...)
}
\arguments{
\item{data}{An interaction data frame.}

\item{...}{The name(s) of the extra attributes; NSE is supported.}
}
\value{
The data frame filtered to the records having the extra attribute.
}
\description{
Interaction records having certain extra attributes
}
\examples{
i <- import_omnipath_interactions(fields = 'extra_attrs')
with_extra_attr(i, Macrophage_type)

}
\seealso{
\itemize{
    \item{\code{\link{extra_attrs}}}
    \item{\code{\link{has_extra_attrs}}}
    \item{\code{\link{extra_attrs_to_cols}}}
}
}
