% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{with_references}
\alias{with_references}
\title{Interactions having references}
\usage{
with_references(data, resources = NULL)
}
\arguments{
\item{data}{An interaction data frame.}

\item{resources}{Character: consider only these resources. If `NULL`,
records with any reference will be accepted.}
}
\value{
A subset of the input interaction data frame.
}
\description{
Interactions having references
}
\examples{
cc <- import_post_translational_interactions(resources = 'CellChatDB')
with_references(cc, 'CellChatDB')

}
