% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{all_uniprots}
\alias{all_uniprots}
\title{A table with all UniProt IDs}
\usage{
all_uniprots(fields = "id", reviewed = TRUE, organism = 9606)
}
\arguments{
\item{fields}{Character vector of fields as defined by UniProt. For
possible values please refer to
\url{https://www.uniprot.org/help/uniprotkb\%5Fcolumn\%5Fnames}}

\item{reviewed}{Retrieve only reviewed (`TRUE`), only unreviewed (`FALSE`)
or both (`NULL`).}

\item{organism}{Integer, NCBI Taxonomy ID of the organism (by default
9606 for human).}
}
\value{
Data frame (tibble) with the requested UniProt entries and fields.
}
\description{
Retrieves a table from UniProt with all proteins for a certain organism.
}
\examples{
human_swissprot_ac <- all_uniprots(fields = 'entry name')
human_swissprot_ac
# # A tibble: 20,396 x 1
#    `Entry name`
#    <chr>
#  1 OR4K3_HUMAN
#  2 O52A1_HUMAN
#  3 O2AG1_HUMAN
#  4 O10S1_HUMAN
#  5 O11G2_HUMAN
# # . with 20,386 more rows
}
