% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma_motif_es}
\alias{bma_motif_es}
\title{BMA motifs from a sequence of edges}
\usage{
bma_motif_es(edge_seq, G, granularity = 2)
}
\arguments{
\item{edge_seq}{An igraph edge sequence.}

\item{G}{An igraph graph object.}

\item{granularity}{Numeric: granularity value.}
}
\value{
Character: BMA motifs as a single string.
}
\description{
These motifs can be added to a BMA canvas.
}
\examples{
interactions <- omnipath(resources = "ARN")
graph <- interaction_graph(interactions)
motifs <- bma_motif_es(igraph::E(graph)[1], graph)

}
