% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{collectri}
\alias{collectri}
\title{TF-target interactions from CollecTRI}
\usage{
collectri(
  resources = NULL,
  organism = 9606L,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = TRUE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources. In case of
CollecTRI this is not desirable for most of the applications. For most of
the other interaction querying functions it is `FALSE` by default.}

\item{...}{Optional additional arguments, passed to
\code{\link{import_transcriptional_interactions}}.}
}
\value{
A dataframe of TF-target interactions.
}
\description{
CollecTRI is a comprehensive resource of transcriptional regulation,
published in 2023, consisting of 14 resources and original literature
curation.
}
\examples{
collectri_grn <- collectri()
collectri_grn

}
\seealso{
\itemize{
    \item{\code{\link{import_transcriptional_interactions}}}
    \item{\code{\link{dorothea}}}
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
