% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{cookie}
\alias{cookie}
\title{Acquire a cookie if necessary}
\usage{
cookie(
  url,
  init_url = NULL,
  post = NULL,
  payload = NULL,
  init_post = NULL,
  init_payload = NULL,
  curl_verbose = FALSE
)
}
\arguments{
\item{url}{Character. URL to download to get the cookie.}

\item{init_url}{Character. An initial URL to download to get the cookie,
before downloading ``url`` with the cookie.}

\item{post}{List: HTTP POST parameters.}

\item{payload}{Data to send as payload.}

\item{init_post}{List: HTTP POST parameters for ``init_url``.}

\item{init_payload}{Data to send as payload with ``init_url``.}

\item{curl_verbose}{Logical. Perform CURL requests in verbose mode for
debugging purposes.}
}
\value{
A list with cache file path, cookies and response headers.
}
\description{
Acquire a cookie if necessary
}
