% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmos_PKN_generation.R
\name{create_PKN_COSMOS}
\alias{create_PKN_COSMOS}
\title{Generating organism-specific PKNs for COSMOS}
\usage{
create_PKN_COSMOS(
  organism,
  translate.genes = FALSE,
  biomart.use.omnipath = TRUE,
  GEM.reactions.map.col = "rxns",
  GEM.metabolites.map.col = "mets",
  GEM.list.params = list(stoich.name = "S", reaction.name = "grRules", lb.name = "lb",
    ub.name = "ub", rev.name = "rev", reaction.ID.name = "rxns", metabolites.ID.name =
    "mets", metabolites.names.name = "metNames", metabolites.fomulas.name =
    "metFormulas", metabolites.inchi.name = "inchis"),
  GEM.degree.mets.threshold = 400,
  stitch.threshold = 700,
  verbose = TRUE
)
}
\arguments{
\item{organism}{Character or integer: an organism (taxon) identifier. 
Supported taxons are 9606 (\emph{Homo sapiens}), 10090 
(\emph{Mus musculus}), and 10116 (\emph{Rattus norvegicus}).}

\item{translate.genes}{Whether translating genes from ENSEMBL into SYMBOL.
Only required when \code{organism == 9606} (\code{FALSE} by default).}

\item{biomart.use.omnipath}{Whether using BioMart information from OmnipathR 
(\code{TRUE} by default).}

\item{GEM.reactions.map.col}{Column of reaction IDs in the GEM
(\code{"rxns"} by default).}

\item{GEM.metabolites.map.col}{Column of reaction IDs in the GEM 
(\code{"mets"} by default).}

\item{GEM.list.params}{List containing the name of the slots where the 
information to construct the PKN is located in the GEM. If a matlab object
is provided, this list parameter should not be modified.}

\item{GEM.degree.mets.threshold}{Degree cutoff used to filter out 
metabolites (400 by default). The objective is to remove cofactors and 
over-promiscuous metabolites.}

\item{stitch.threshold}{Confidence cutoff used for STITCH connections 
(700 by default).}

\item{verbose}{Whether showing messages during execution (\code{TRUE} by 
default).}
}
\value{
List of 4 elements containing the necessary information for COSMOS to 
  run: causal PKN, mapping data frame for metabolites from GEM, 
  reaction-to-gene data frame from GEM, and mapping data frame for reactions 
  from GEM.
}
\description{
This function generates the prior knowledge network (PKN) needed to run 
COSMOS using information from different resources through the \pkg{OmnipathR} 
R package. Particularly, \code{create_PKN_COSMOS} will obtain: 
\itemize{ \item Genome-scale metabolic 
model (GEM) of the required organism from Wang et al., 2021. 
\item Interaction network of 
chemicals and proteins from STITCH (\url{http://stitch.embl.de/}) for the 
required organism. \item Protein-protein interactions from Omnipath (Türei 
et al., 2021) for the required organism} With these three pieces of 
information, the function will generate the required causal network for 
COSMOS to run.
}
\examples{
\dontrun{
  human.PKN.COSMOS <- create_PKN_COSMOS(organism = 9606)
}

}
\references{
Wang H, Robinson JL, Kocabas P, Gustafsson J, Anton M, 
   Cholley PE, et al. Genome-scale metabolic network reconstruction of model 
   animals as a platform for translational research. Proceedings of the 
   National Academy of Sciences. 2021 Jul 27;118(30):e2102344118.
   
   Türei D, Valdeolivas A, Gul L, Palacio‐Escat N, Klein M, Ivanova O, et al.
    Integrated intra‐ and intercellular signaling knowledge for multicellular 
    omics analysis. Molecular Systems Biology. 2021 Mar;17(3):e9923.
}
