% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{dorothea}
\alias{dorothea}
\alias{import_tfregulons_interactions}
\alias{import_dorothea_interactions}
\title{TF-target interactions from DoRothEA}
\usage{
dorothea(
  resources = NULL,
  organism = 9606,
  dorothea_levels = c("A", "B"),
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = TRUE,
  genesymbol_resource = NULL,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Character or integer: Name or NCBI Taxonomy ID of one or
organisms. The web service currently provides interactions for
human, mouse and rat. For other organisms, the data will be translated
by orthologous gene pairs from human. In this case, only one organism
can be provided.}

\item{dorothea_levels}{Vector detailing the confidence levels of the
interactions to be downloaded. In dorothea, every TF-target interaction
has a confidence score ranging from A to E, being A the most reliable
interactions.
By default we take A and B level interactions (\code{c(A, B)}).
It is to note that E interactions are not available in OmnipathR.}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources. In case of
DoRothEA this is not desirable for most of the applications. For most of
the other interaction querying functions it is `FALSE` by default.}

\item{genesymbol_resource}{Character: either "uniprot" or "ensembl". The
former leaves intact the gene symbols returned by the web service,
originally set from UniProt. The latter updates the gene symbols from
Ensembl, which uses a slightly different gene symbol standard. In this
case a few records will be duplicated, where Ensembl provides ambiguous
translation.}

\item{...}{optional additional arguments}
}
\value{
A dataframe of TF-target interactions from DoRothEA
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=dorothea}
which contains transcription factor (TF)-target interactions from DoRothEA
\url{https://github.com/saezlab/DoRothEA}
DoRothEA is a comprehensive resource of transcriptional regulation,
consisting of 16 original resources, in silico TFBS prediction, gene
expression signatures and ChIP-Seq binding site analysis.
}
\examples{
dorothea_grn <- dorothea(
    resources = c('DoRothEA', 'ARACNe-GTEx_DoRothEA'),
    organism = 9606,
    dorothea_levels = c('A', 'B', 'C')
)
dorothea_grn

}
\seealso{
\itemize{
    \item{\code{\link{collectri}}}
    \item{\code{\link{import_transcriptional_interactions}}}
    \item{\code{\link{import_tf_target_interactions}}}
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
