% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{.get_interaction_databases}
\alias{.get_interaction_databases}
\title{Get the different interaction databases}
\usage{
.get_interaction_databases()
}
\value{
character vector with the names of the interaction databases
}
\description{
get the names of the databases from \url{http://omnipath.org/interactions}
}
\examples{
.get_interaction_databases()
}
\seealso{
\code{\link{import_AllInteractions}, 
\link{import_Omnipath_Interactions}, \link{import_PathwayExtra_Interactions},
\link{import_KinaseExtra_Interactions}, 
\link{import_LigrecExtra_Interactions},
\link{import_miRNAtarget_Interactions}, 
\link{import_TFregulons_Interactions}}
}
