% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{.get_ptms_databases}
\alias{.get_ptms_databases}
\title{Get Post-translational modification (PTMs) databases}
\usage{
.get_ptms_databases()
}
\value{
character vector with the names of the PTMs databases
}
\description{
get the names of the different databases available for ptms databases 
\url{http://omnipath.org/ptms}
}
\examples{
.get_ptms_databases()
}
\seealso{
\code{\link{import_Omnipath_PTMS}}
}
