% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_annotation_resources}
\alias{get_annotation_resources}
\alias{get_annotation_databases}
\title{Get the resources available in the annotations database of OmniPath}
\usage{
get_annotation_resources(dataset = NULL, ...)

get_annotation_databases(dataset = NULL, ...)
}
\arguments{
\item{dataset}{ignored for this query type}

\item{...}{Passed to \code{get_annotation_resources}.}
}
\value{
character vector with the names of the annotation resources
}
\description{
get the names of the resources from \url{http://omnipath.org/annotations}
}
\examples{
get_annotation_resources()

}
\seealso{
\code{\link{get_resources},
\link{import_omnipath_annotations}}
}
