% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_complex_resources}
\alias{get_complex_resources}
\alias{get_complexes_databases}
\title{Retrieve a list of complex resources available in Omnipath}
\usage{
get_complex_resources(dataset = NULL)

get_complexes_databases(...)
}
\arguments{
\item{dataset}{ignored for this query type}

\item{...}{Passed to \code{import_omnipath_enzsub}.}
}
\value{
character vector with the names of the databases
}
\description{
get the names of the resources from \url{http://omnipath.org/complexes}
}
\examples{
get_complex_resources()

}
\seealso{
\code{\link{get_resources},
\link{import_omnipath_complexes}}
}
