% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{get_interaction_resources}
\alias{get_interaction_resources}
\alias{get_interaction_databases}
\title{Retrieve a list of interaction resources available in Omnipath}
\usage{
get_interaction_resources(dataset = NULL)
}
\arguments{
\item{dataset}{a dataset within the interactions query type. Currently
available datasets are `omnipath`, `kinaseextra`, `pathwayextra`,
`ligrecextra`, `dorothea`, `tf_target`, `tf_mirna`, `mirnatarget` and
`lncrna_mrna`}
}
\value{
character vector with the names of the interaction databases
}
\description{
Gets the names of the resources from
\url{https://omnipath.org/interactions}.
}
\examples{
get_interaction_resources()

}
\seealso{
\code{\link{get_resources},
\link{import_all_interactions},
\link{import_omnipath_interactions}, \link{import_pathwayextra_interactions},
\link{import_kinaseextra_interactions},
\link{import_ligrecextra_interactions},
\link{import_mirnatarget_interactions},
\link{import_dorothea_interactions}}
}
