% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{get_intercell_categories}
\alias{get_intercell_categories}
\title{Retrieves a list of categories from the intercell database of OmniPath}
\usage{
get_intercell_categories()
}
\value{
character vector with the different intercell categories
}
\description{
Retrieves a list of categories from \url{https://omnipath.org/intercell}.
}
\examples{
\donttest{
get_intercell_categories()
}

}
\seealso{
\code{\link{import_omnipath_intercell},
    \link{get_intercell_generic_categories}}
}
