% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_intercell_categories}
\alias{get_intercell_categories}
\title{Get the different intercell categories described in Omnipath}
\usage{
get_intercell_categories()
}
\value{
character vector with the different intercell categories
}
\description{
get the names of the categories from \url{http://omnipath.org/intercell}
}
\examples{
get_intercell_categories()
}
\seealso{
\code{\link{import_Omnipath_intercell}}
}
