% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{get_intercell_generic_categories}
\alias{get_intercell_generic_categories}
\alias{get_intercell_classes}
\title{Retrieve a list of the generic categories in the intercell database
of OmniPath}
\usage{
get_intercell_generic_categories()

get_intercell_classes()
}
\value{
character vector with the different intercell main classes
}
\description{
get the names of the generic categories from
\url{http://omnipath.org/intercell}
}
\examples{
get_intercell_generic_categories()
}
\seealso{
\code{\link{import_omnipath_intercell},
\link{get_intercell_categories}}
}
