% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{get_intercell_generic_categories}
\alias{get_intercell_generic_categories}
\alias{get_intercell_classes}
\title{Retrieves a list of the generic categories in the intercell database
of OmniPath}
\usage{
get_intercell_generic_categories()
}
\value{
character vector with the different intercell main classes
}
\description{
Retrieves a list of the generic categories from
\url{https://omnipath.org/intercell}.
}
\examples{
\donttest{
get_intercell_generic_categories()
}

}
\seealso{
\itemize{
    \item{\code{\link{import_omnipath_intercell}}}
    \item{\code{\link{get_intercell_categories}}}
}
}
