% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{get_reactome_pathway_relations}
\alias{get_reactome_pathway_relations}
\title{Reactome pathway relations}
\usage{
get_reactome_pathway_relations(species = NULL)
}
\arguments{
\item{species}{Character: species code (e.g. "HSA") or full name
(e.g. "Homo sapiens"). If `NULL`, returns all species.}
}
\value{
A tibble with columns `Parent`, `Child`.
}
\description{
Retrieves parent-child relationships between Reactome pathways and
optionally filters by species.
}
\examples{
.slow_doctest()
\dontrun{
relations <- get_reactome_pathway_relations("HSA")
head(relations)
}
}
