% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signed_ptms.R
\name{get_signed_ptms}
\alias{get_signed_ptms}
\title{get signs for ptms interactions}
\usage{
get_signed_ptms(
  ptms = import_omnipath_enzsub(),
  interactions = import_omnipath_interactions()
)
}
\arguments{
\item{ptms}{ptms data frame generated by \code{\link{import_omnipath_enzsub}}}

\item{interactions}{interaction data frame generated by
\code{\link{import_omnipath_interactions}}}
}
\value{
data.frame of ptms with is_inhibition and is_stimulation columns
}
\description{
ptms data does not contain sign (activation/inhibition), we generate this
information based on the interaction network
}
\examples{
ptms = import_omnipath_enzsub(resources=c("PhosphoSite", "SIGNOR"))
interactions = import_omnipath_interactions()
ptms = get_signed_ptms(ptms,interactions)
}
\seealso{
\code{\link{import_omnipath_enzsub}}
\code{\link{import_omnipath_interactions}}
}
