% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzsub_signs.R
\name{get_signed_ptms}
\alias{get_signed_ptms}
\title{Signs for enzyme-substrate interactions}
\usage{
get_signed_ptms(
  enzsub = import_omnipath_enzsub(),
  interactions = import_omnipath_interactions()
)
}
\arguments{
\item{interactions}{interaction data frame generated by
\code{\link{import_omnipath_interactions}}}

\item{enzsib}{Enzyme-substrate data frame generated by
\code{\link{import_omnipath_enzsub}}}
}
\value{
Data frame of enzyme-substrate relationships with is_inhibition
    and is_stimulation columns.
}
\description{
Enzyme-substrate data does not contain sign (activation/inhibition), we
generate this information based on the interaction network.
}
\examples{
enzsub <- import_omnipath_enzsub(resources = c('PhosphoSite', 'SIGNOR'))
interactions <- import_omnipath_interactions()
enzsub <- get_signed_ptms(enzsub, interactions)

}
\seealso{
\itemize{
    \item{\code{\link{import_omnipath_enzsub}}}
    \item{\code{\link{import_omnipath_interactions}}}
}
}
