% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go_annot_slim}
\alias{go_annot_slim}
\title{GO slim gene annotations}
\usage{
go_annot_slim(organism = "human", slim = "generic", aspects = c("C", "F", "P"))
}
\arguments{
\item{organism}{Character: either "chicken", "cow", "dog", "human", "pig"
or "uniprot_all".}

\item{aspects}{Character vector with some of the following elements:
"C" (cellular component), "F" (molecular function) and "P" (biological
process). Gene Ontology is three separate ontologies called as three
aspects. By this parameter you can control which aspects to include
in the output.}

\item{subset}{Character: the GO subset (GO slim) name. Available GO
slims are: "agr" (Alliance for Genomics Resources), "generic",
"aspergillus", "candida", "drosophila", "chembl", "metagenomic",
"mouse", "plant", "pir" (Protein Information Resource), "pombe"
and "yeast".}
}
\value{
A tibble (data frame) of genes annotated with ontology terms in
    in the GO slim (subset).
}
\description{
GO slims are subsets of the full GO which "give a broad overview of the
ontology content without the detail of the specific fine grained terms".
In order to annotate genes with GO slim terms, we take the annotations
and search all ancestors of the terms up to the root of the ontology
tree. From the ancestors we select the terms which are part of the slim
subset.
}
