% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_intercell_network}
\alias{import_intercell_network}
\title{Imports an intercellular network combining annotations and interactions}
\usage{
import_intercell_network(
  interactions_param = list(),
  transmitter_param = list(),
  receiver_param = list(),
  resources = NULL,
  entity_types = NULL,
  ligand_receptor = FALSE,
  high_confidence = FALSE,
  simplify = FALSE,
  consensus_percentile = NULL,
  ...
)
}
\arguments{
\item{interactions_param}{a list with arguments for an interactions query:
\code{\link{import_omnipath_interactions}},
\code{\link{import_pathwayextra_interactions}},
\code{\link{import_kinaseextra_interactions}},
\code{\link{import_ligrecextra_interactions}}}

\item{transmitter_param}{a list with arguments for
\code{\link{import_omnipath_intercell}}, to define the transmitter side
of intercellular connections}

\item{receiver_param}{a list with arguments for
\code{\link{import_omnipath_intercell}}, to define the receiver side
of intercellular connections}

\item{resources}{A character vector of resources to be applied to
both the interactions and the annotations. For example, \code{resources
= 'CellChatDB'} will download the transmitters and receivers defined by
CellChatDB, connected by connections from CellChatDB.}

\item{entity_types}{Character, possible values are "protein", "complex" or
both.}

\item{ligand_receptor}{Logical. If TRUE, only *ligand* and *receptor*
annotations will be used instead of the more generic *transmitter* and
*receiver* categories.}

\item{high_confidence}{Logical: shortcut to do some filtering in order to
include only higher confidence interactions. The intercell database
of OmniPath covers a very broad range of possible ways of cell to cell
communication, and the pieces of information, such as localization,
topology, function and interaction, are combined from many, often
independent sources. This unavoidably result some weird and unexpected
combinations which are false positives in the context of intercellular
communication. This option sets some minimum criteria to remove most
(but definitely not all!) of the wrong connections. These criteria
are the followings: 1) the receiver must be plasma membrane
transmembrane; 2) the curation effort for interactions must be larger
than one; 3) the consensus score for annotations must be larger than
one. 4) the transmitter must be secreted or exposed on the plasma
membrane. These are very relaxed criteria, you can always tune them
to be more stringent by filtering manually.}

\item{simplify}{Logical: keep only the most often used columns. This
function combines a network data frame with two copies of the
intercell annotation data frames, all of them already having quite
some columns. With this option we keep only the names of the
interacting pair, their intercellular communication roles, and the
minimal information of the origin of both the interaction and
the annotations.}

\item{consensus_percentile}{Numeric: a percentile cut off for the consensus
score of generic categories in intercell annotations. The consensus
score is the number of resources supporting the classification of an
entity into a category based on combined information of many resources.
Here you can apply a cut-off, keeping only the annotations supported
by a higher number of resources than a certain percentile of each
category. If \code{NULL} no filtering will be performed. The value is
either in the 0-1 range, or will be divided by 100 if graeter than 1.
The percentiles will be calculated against the generic composite
categories and then will be applied to their resource specific
annotations and specific child categories.}

\item{...}{If \code{simplify} is \code{TRUE}, additional column
names can be passed here to \code{dplyr::select} on the final
data frame. Otherwise ignored.}
}
\value{
A dataframe containing information about protein-protein
interactions and the inter-cellular roles of the protiens involved in those
interactions.
}
\description{
Imports an intercellular network by mapping intercellular annotations
and protein interactions. First imports a network of protein-protein
interactions. Then, it retrieves annotations about the proteins
intercellular communication roles, once for the transmitter (delivering
information from the expressing cell) and second, the receiver (receiving
signal and relaying it towards the expressing cell) side. These 3 queries
can be customized by providing parameters in lists which will be passed to
the respective methods (\code{\link{import_omnipath_interactions}} for
the network and \code{\link{import_omnipath_intercell}} for the
annotations). Finally the 3 data frames combined in a way that the source
proteins in each interaction annotated by the transmitter, and the target
proteins by the receiver categories. If undirected interactions present
(these are disabled by default) they will be duplicated, i.e. both
partners can be both receiver and transmitter.
}
\examples{
intercell_network <- import_intercell_network(
   interactions_param = list(datasets = 'ligrecextra'),
   receiver_param = list(categories = c('receptor', 'transporter')),
   transmitter_param = list(categories = c('ligand', 'secreted_enzyme'))
)

}
\seealso{
\itemize{
    \item{\code{\link{get_intercell_categories}}}
    \item{\code{\link{get_intercell_generic_categories}}}
    \item{\code{\link{import_omnipath_intercell}}}
    \item{\code{\link{import_omnipath_interactions}}}
    \item{\code{\link{import_pathwayextra_interactions}}}
    \item{\code{\link{import_kinaseextra_interactions}}}
    \item{\code{\link{import_ligrecextra_interactions}}}
}
}
