% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_intercell_network}
\alias{import_intercell_network}
\title{Imports an intercellular network combining annotations and interactions}
\usage{
import_intercell_network(
  cache_file = NULL,
  interactions_param = list(),
  transmitter_param = list(),
  receiver_param = list()
)
}
\arguments{
\item{cache_file}{path to an earlier data file; if exists, will be loaded
as it is, the further arguments have no effect; if does not exists, the
result will be dumped into this file.}

\item{interactions_param}{a list with arguments for an interactions query: 
\code{\link{import_omnipath_interactions}, 
\link{import_pathwayextra_interactions},
\link{import_kinaseextra_interactions},
\link{import_ligrecextra_interactions}}}

\item{transmitter_param}{a list with arguments for
\code{\link{import_omnipath_intercell}}, to define the transmitter side
of intercellular connections}

\item{receiver_param}{a list with arguments for
\code{\link{import_omnipath_intercell}}, to define the receiver side
of intercellular connections}
}
\value{
A dataframe containing information about protein-protein
interactions and the inter-cellular roles of the protiens involved in those
interactions.
}
\description{
Imports an intercellular network by mapping intercellular annotations
and protein interactions. First imports a network of protein-protein
interactions. Then, it retrieves annotations about the proteins
intercellular communication roles, once for the transmitter (delivering
information from the expressing cell) and second, the receiver (receiving
signal and relaying it towards the expressing cell) side. These 3 queries
can be customized by providing parameters in lists which will be passed to
the respective methods (\code{\link{import_omnipath_interactions}} for
the network and \code{\link{import_omnipath_intercell}} for the
annotations). Finally the 3 data frames combined in a way that the source
proteins in each interaction annotated by the transmitter, and the target
proteins by the receiver categories. If undirected interactions present
(these are disabled by default) they will be duplicated, i.e. both
partners can be both receiver and transmitter.
If a cache file provided, its content will be returned without any further
filtering.
}
\examples{
intercellNetwork <- import_intercell_network(
   interactions_param = list(datasets = 'ligrecextra'),
   receiver_param = list(categories = c('receptor', 'transporter')),
   transmitter_param = list(categories = c('ligand', 'secreted_enzyme')))

}
\seealso{
\code{\link{get_intercell_categories},
\link{get_intercell_generic_categories}, 
\link{import_omnipath_intercell},
\link{import_omnipath_interactions}, 
\link{import_pathwayextra_interactions},
\link{import_kinaseextra_interactions}, 
\link{import_ligrecextra_interactions}}
}
