% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_miRNAtarget_Interactions}
\alias{import_miRNAtarget_Interactions}
\title{Imports from Omnipath webservice the interactions from miRNAtarget dataset}
\usage{
import_miRNAtarget_Interactions(from_cache_file = NULL,
  filter_databases = .get_interaction_databases())
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are 
removed. See \code{\link{.get_interaction_databases}} for more information.}
}
\value{
A dataframe containing miRNA-mRNA and TF-miRNA interactions
}
\description{
Imports the dataset from: 
\url{http://omnipathdb.org/interactions?datasets=mirnatarget}, 
which contains miRNA-mRNA and TF-miRNA interactions
}
\examples{
interactions <- 
  import_miRNAtarget_Interactions(filter_databases=c("miRTarBase",
  "miRecords"))
}
\seealso{
\code{\link{.get_interaction_databases}, 
  \link{import_AllInteractions}}
}
