% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_Omnipath_annotations}
\alias{import_Omnipath_annotations}
\title{Import Omnipath Annotations}
\usage{
import_Omnipath_annotations(
  from_cache_file = NULL,
  select_genes = NULL,
  filter_databases = get_annotation_databases()
)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{select_genes}{vector containing the genes for whom annotations will be
retrieved (hgnc format). It is also possible to donwload complexes 
annotations. To do so, write "COMPLEX:" right before the genesymbols of
the genes integrating the complex. Check the vignette for examples.}

\item{filter_databases}{annotations not reported in these databases are 
removed. See \code{\link{get_annotation_databases}} for more information.}
}
\value{
A data.frame containing different gene/complex annotations
}
\description{
imports the annotations stored in Omnipath database from 
\url{http://omnipathdb.org/annotations}
}
\examples{
annotations = import_Omnipath_annotations(select_genes=c("TP53","LMNA"),
     filter_databases=c("HPA_subcellular"))
}
\seealso{
\code{\link{get_annotation_databases}}
}
