% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_omnipath_complexes}
\alias{import_omnipath_complexes}
\alias{import_Omnipath_complexes}
\alias{import_OmniPath_complexes}
\title{Imports protein complexes from OmniPath}
\usage{
import_omnipath_complexes(cache_file = NULL, resources = NULL, ...)
}
\arguments{
\item{cache_file}{path to an earlier data file}

\item{resources}{complexes not reported in these databases are
removed. See \code{\link{get_complexes_databases}} for more information.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing information about complexes
}
\description{
Imports the complexes stored in Omnipath database from
\url{https://omnipathdb.org/complexes}.
}
\examples{
\donttest{
complexes = import_omnipath_complexes(
    resources = c('CORUM', 'hu.MAP')
)
}

}
\seealso{
\code{\link{get_complexes_databases}}
}
