% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_omnipath_enzsub}
\alias{import_omnipath_enzsub}
\alias{import_Omnipath_PTMS}
\alias{import_OmniPath_PTMS}
\title{Imports enzyme-substrate relationships from OmniPath}
\usage{
import_omnipath_enzsub(
  cache_file = NULL,
  resources = NULL,
  organism = 9606,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  ...
)
}
\arguments{
\item{cache_file}{path to an earlier data file}

\item{resources}{PTMs not reported in these databases are
removed. See \code{\link{get_ptms_databases}} for more information}

\item{organism}{PTMs are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{...}{Optional additional arguments.}
}
\value{
A data frame containing the information about ptms
}
\description{
Imports the enzyme-substrate (more exactly, enzyme-PTM) relationship
database from \url{https://omnipathdb.org/enzsub}
}
\examples{
\donttest{
ptms = import_omnipath_enzsub(
    resources = c('PhosphoSite', 'SIGNOR'),
    organism = 9606
)
}

}
\seealso{
\itemize{
    \item{\code{\link{get_ptms_databases}}}
    \item{\code{\link{import_omnipath_interactions}}}
}
}
