% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_Omnipath_Interactions}
\alias{import_Omnipath_Interactions}
\title{Import Omnipath interaction database}
\usage{
import_Omnipath_Interactions(
  from_cache_file = NULL,
  filter_databases = get_interaction_databases(),
  select_organism = 9606
)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{filter_databases}{interactions not reported in these databases are 
removed. See \code{\link{get_interaction_databases}} for more information.}

\item{select_organism}{Interactions are available for human, mouse and rat. 
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}
}
\value{
A dataframe containing information about protein-protein interactions
}
\description{
imports the database from \url{http://omnipathdb.org/interactions}, which 
contains only interactions with references. These interactions are the 
original ones from the first Omnipath version.
}
\examples{
interactions = import_Omnipath_Interactions(filter_databases=c("SignaLink3"),
               select_organism = 9606)
               
}
\seealso{
\code{\link{get_interaction_databases}, 
  \link{import_AllInteractions}}
}
