% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{import_omnipath_interactions}
\alias{import_omnipath_interactions}
\title{Imports interactions from the `omnipath` dataset of Omnipath}
\usage{
import_omnipath_interactions(
  resources = NULL,
  organism = "human",
  datasets = "omnipath",
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Character or integer: Name or NCBI Taxonomy ID of one or
organisms. The web service currently provides interactions for
human, mouse and rat. For other organisms, the data will be translated
by orthologous gene pairs from human. In this case, only one organism
can be provided.}

\item{datasets}{Names of the interaction datasets to download: omnipath
(by default). Other possiblites are: pathwayextra, kinaseextra,
ligrecextra, dorothea,tf_target, mirnatarget, tf_mirna, lncrna_mrna.
The user can select multiple datasets as for example: c('omnipath',
'pathwayextra', 'kinaseextra')}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{...}{optional additional arguments}
}
\value{
A dataframe of protein-protein interactions
}
\description{
Imports the database from \url{https://omnipathdb.org/interactions}, which
contains only interactions supported by literature references.
This part of the interaction database compiled a similar way as it has
been presented in the first paper describing OmniPath (Turei et al. 2016).
}
\examples{
interactions = import_omnipath_interactions(
    resources = c('SignaLink3'),
    organism = 9606
)

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
