% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_omnipath_interactions}
\alias{import_omnipath_interactions}
\alias{import_Omnipath_Interactions}
\alias{import_OmniPath_Interactions}
\title{Imports interactions from the `omnipath` dataset of Omnipath}
\usage{
import_omnipath_interactions(
  resources = NULL,
  organism = 9606,
  datasets = "omnipath",
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{datasets}{Names of the interaction datasts to download: omnipath 
(by default). Other possiblites are: pathwayextra, kinaseextra,
ligrecextra, dorothea,tf_target, mirnatarget, tf_mirna, lncrna_mrna.
The user can select multiple datasets as for example: c('omnipath',
'pathwayextra', 'kinaseextra')}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{...}{optional additional arguments}
}
\value{
A dataframe of protein-protein interactions
}
\description{
Imports the database from \url{https://omnipathdb.org/interactions}, which
contains only interactions supported by literature references.
This part of the interaction database compiled a similar way as it has
been presented in the first paper describing OmniPath (Turei et al. 2016).
}
\examples{
interactions = import_omnipath_interactions(
    resources = c('SignaLink3'),
    organism = 9606
)

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
