% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_Omnipath_intercell}
\alias{import_Omnipath_intercell}
\title{Import Omnipath Intercell Data}
\usage{
import_Omnipath_intercell(
  from_cache_file = NULL,
  select_categories = get_intercell_categories(),
  select_classes = get_intercell_classes()
)
}
\arguments{
\item{from_cache_file}{path to an earlier data file}

\item{select_categories}{vector containing the categories to be retrieved.
All the genes belonging to those categories will be returned. For furter 
information about the categories see \code{\link{get_intercell_categories}}}

\item{select_classes}{vector containing the main classes to be retrieved.
All the genes belonging to those classes will be returned. For furter 
information about the main classes see \code{\link{get_intercell_classes}}}
}
\value{
A dataframe cotaining information about roles in inter-cellular
signaling.
}
\description{
imports the intercell data stored in Omnipath database from 
\url{http://omnipathdb.org/intercell}. Intercell provides 
information on the roles in inter-cellular signaling. E.g. if a protein is 
a ligand, a receptor, an extracellular matrix (ECM) component, etc.
}
\examples{
intercell = import_Omnipath_intercell(select_categories=c("ecm"))
}
\seealso{
\code{\link{get_intercell_categories}}
}
