% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_omnipath_intercell}
\alias{import_omnipath_intercell}
\alias{import_Omnipath_intercell}
\alias{import_OmniPath_intercell}
\title{Imports OmniPath intercell annotations}
\usage{
import_omnipath_intercell(
  categories = NULL,
  resources = NULL,
  parent = NULL,
  scope = NULL,
  aspect = NULL,
  source = NULL,
  transmitter = NULL,
  receiver = NULL,
  secreted = NULL,
  plasma_membrane_peripheral = NULL,
  plasma_membrane_transmembrane = NULL,
  proteins = NULL,
  topology = NULL,
  causality = NULL,
  ...
)
}
\arguments{
\item{categories}{vector containing the categories to be retrieved.
All the genes belonging to those categories will be returned. For further
information about the categories see \code{\link{get_intercell_categories}}}

\item{resources}{limit the query to certain resources; see the available
resources by \code{\link{get_intercell_resources}}}

\item{parent}{vector containing the parent classes to be retrieved.
All the genes belonging to those classes will be returned. For furter
information about the main classes see
\code{\link{get_intercell_categories}}}

\item{scope}{either `specific` or `generic`}

\item{aspect}{either `locational` or `functional`}

\item{source}{either `resource_specific` or `composite`}

\item{transmitter}{logical, include only transmitters i.e. proteins
delivering signal from a cell to
its environment}

\item{receiver}{logical, include only receivers i.e. proteins delivering
signal to the cell from its environment}

\item{secreted}{logical, include only secreted proteins}

\item{plasma_membrane_peripheral}{logical, include only plasma membrane
peripheral membrane proteins}

\item{plasma_membrane_transmembrane}{logical, include only plasma membrane
transmembrane proteins}

\item{proteins}{limit the query to certain proteins}

\item{topology}{topology categories: one or more of `secreted` (sec),
`plasma_membrane_peripheral` (pmp), `plasma_membrane_transmembrane` (pmtm)
(both short or long notation can be used)}

\item{causality}{`transmitter` (trans), `receiver` (rec) or `both` (both
short or long notation can be used)}

\item{...}{Additional optional arguments}
}
\value{
A dataframe cotaining information about roles in intercellular
signaling.
}
\description{
Imports the OmniPath intercellular communication role annotation database
from \url{https://omnipathdb.org/intercell}. It provides information
on the roles in inter-cellular signaling. E.g. if a protein is
a ligand, a receptor, an extracellular matrix (ECM) component, etc.
}
\examples{
intercell <- import_omnipath_intercell(categories = 'ecm')

}
\seealso{
\itemize{
    \item{\code{\link{get_intercell_categories}}}
    \item{\code{\link{get_intercell_generic_categories}}}
    \item{\code{\link{import_intercell_network}}}
}
}
