% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{import_pathwayextra_interactions}
\alias{import_pathwayextra_interactions}
\alias{import_PathwayExtra_Interactions}
\title{Imports interactions from the `pathway extra` dataset of Omnipath}
\usage{
import_pathwayextra_interactions(
  resources = NULL,
  organism = 9606,
  fields = NULL,
  default_fields = TRUE,
  references_by_resource = TRUE,
  exclude = NULL,
  strict_evidences = FALSE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose one of those: 9606 human (default), 10116 rat or 10090 Mouse.}

\item{fields}{The user can define here the fields to be added. If used, set
the next argument, `default_fields`, to FALSE.}

\item{default_fields}{whether to include the default fields (columns) for
the query type. If FALSE, only the fields defined by the user in the
`fields` argument will be added.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{strict_evidences}{Logical: restrict the evidences to the queried
datasets and resources. If set to FALSE, the directions and effect signs
and references might be based on other datasets and resources.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing activity flow interactions between proteins
without literature reference
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=pathwayextra},
which contains activity flow interactions without literature reference.
The activity flow interactions supported by literature references
are part of the `omnipath` dataset.
}
\examples{
interactions <-
    import_pathwayextra_interactions(
        resources = c('BioGRID', 'IntAct'),
        organism = 9606
    )

}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
