% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Omnipath.R
\name{import_post_translational_interactions}
\alias{import_post_translational_interactions}
\title{Imports all post-translational interactions from OmniPath}
\usage{
import_post_translational_interactions(
  resources = NULL,
  organism = 9606,
  exclude = NULL,
  references_by_resource = TRUE,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_databases}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{exclude}{datasets to exclude}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{...}{optional additional arguments}
}
\value{
A dataframe containing post-translational interactions
}
\description{
Imports the dataset from all post-translational datasets of OmniPath
}
\examples{
interactions <-
    import_post_translational_interactions(
        resources = c('BioGRID')
    )

}
\seealso{
\code{\link{get_interaction_databases},
  \link{import_all_interactions}}
}
