% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{import_transcriptional_interactions}
\alias{import_transcriptional_interactions}
\title{Imports all TF-target interactions from OmniPath}
\usage{
import_transcriptional_interactions(
  resources = NULL,
  organism = 9606,
  dorothea_levels = c("A", "B"),
  references_by_resource = TRUE,
  exclude = NULL,
  ...
)
}
\arguments{
\item{resources}{interactions not reported in these databases are
removed. See \code{\link{get_interaction_resources}} for more information.}

\item{organism}{Interactions are available for human, mouse and rat.
Choose among: 9606 human (default), 10116 rat and 10090 Mouse}

\item{dorothea_levels}{Vector detailing the confidence levels of the
interactions to be downloaded. In dorothea, every TF-target interaction
has a confidence score ranging from A to E, being A the most reliable
interactions.
By default we take A and B level interactions (\code{c(A, B)}).
It is to note that E interactions are not available in OmnipathR.}

\item{references_by_resource}{if FALSE, removes the resource name prefixes
from the references (PubMed IDs); this way the information which reference
comes from which resource will be lost and the PubMed IDs will be unique.}

\item{exclude}{Character: datasets or resources to exclude.}

\item{...}{Optional additional arguments.}

\item{datasets}{Character vector with the}
}
\value{
A dataframe containing TF-target interactions.
}
\description{
Imports the dataset from:
\url{https://omnipathdb.org/interactions?datasets=tf_target,dorothea},
which contains transcription factor-target protein coding gene
interactions.
}
\examples{
interactions <-
    import_transcriptional_interactions(
        resources = c('PAZAR', 'ORegAnno', 'DoRothEA')
    )


}
\seealso{
\itemize{
    \item{\code{\link{get_interaction_resources}}}
    \item{\code{\link{import_all_interactions}}}
    \item{\code{\link{interaction_graph}}}
    \item{\code{\link{print_interactions}}}
}
}
