% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{interaction_graph}
\alias{interaction_graph}
\title{Build Omnipath interaction graph}
\usage{
interaction_graph(interactions = interactions)
}
\arguments{
\item{interactions}{data.frame created by \itemize{
\item{\code{\link{import_omnipath_enzsub}}}
\item{\code{\link{import_omnipath_interactions}}}
\item{\code{\link{import_pathwayextra_interactions}}}
\item{\code{\link{import_kinaseextra_interactions}}}
\item{\code{\link{import_ligrecextra_interactions}}}
\item{\code{\link{import_post_translationsl_interactions}}}
\item{\code{\link{import_dorothea_interactions}}}
\item{\code{\link{import_tf_target_interactions}}}
\item{\code{\link{import_transcriptional_interactions}}}
\item{\code{\link{import_mirnatarget_interactions}}}
\item{\code{\link{import_all_interactions}}}}}
}
\value{
An igraph graph object.
}
\description{
Transforms the interactions data frame to an igraph graph object.
}
\examples{
interactions <- import_omnipath_interactions(resources = c('SignaLink3'))
g <- interaction_graph(interactions)

}
\seealso{
\itemize{
    \item{\code{\link{import_omnipath_interactions}}}
    \item{\code{\link{import_pathwayextra_interactions}}}
    \item{\code{\link{import_kinaseextra_interactions}}}
    \item{\code{\link{import_ligrecextra_interactions}}}
    \item{\code{\link{import_dorothea_interactions}}}
    \item{\code{\link{import_mirnatarget_interactions}}}
    \item{\code{\link{import_all_interactions}}}
}
}
