% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2graph.R
\name{interaction_graph}
\alias{interaction_graph}
\title{Build Omnipath interaction graph}
\usage{
interaction_graph(interactions)
}
\arguments{
\item{interactions}{object created by \code{\link{import_Omnipath_Interactions}}}
}
\value{
igraph object
}
\description{
transforms the interactions data.frame to igraph object
}
\examples{
interactions = import_Omnipath_Interactions(filter_databases=c("SignaLink3","PhosphoSite", "Signor"))
OPI_g = interaction_graph(interactions = interactions )
}
