% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2graph.R
\name{interaction_graph}
\alias{interaction_graph}
\title{Build Omnipath interaction graph}
\usage{
interaction_graph(interactions = interactions)
}
\arguments{
\item{interactions}{data.frame created by 
\code{\link{import_Omnipath_Interactions}},
\code{\link{import_PathwayExtra_Interactions}}, 
\code{\link{import_KinaseExtra_Interactions}},
\code{\link{import_LigrecExtra_Interactions}}, 
\code{\link{import_TFregulons_Interactions}},
\code{\link{import_miRNAtarget_Interactions}} or 
\code{\link{import_AllInteractions}}}
}
\value{
An igraph object
}
\description{
transforms the interactions data.frame to an igraph object
}
\examples{
interactions = import_Omnipath_Interactions(filter_databases=c("SignaLink3"))
OPI_g = interaction_graph(interactions)
}
\seealso{
\code{\link{import_Omnipath_Interactions}},
\code{\link{import_PathwayExtra_Interactions}}, 
\code{\link{import_KinaseExtra_Interactions}},
\code{\link{import_LigrecExtra_Interactions}}, 
\code{\link{import_TFregulons_Interactions}},
\code{\link{import_miRNAtarget_Interactions}} or 
\code{\link{import_AllInteractions}}
}
