% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{interaction_graph}
\alias{interaction_graph}
\title{Build Omnipath interaction graph}
\usage{
interaction_graph(interactions = interactions)
}
\arguments{
\item{interactions}{data.frame created by
\code{\link{import_omnipath_interactions}},
\code{\link{import_pathwayextra_interactions}},
\code{\link{import_kinaseextra_interactions}},
\code{\link{import_ligrecextra_interactions}},
\code{\link{import_dorothea_interactions}},
\code{\link{import_mirnatarget_interactions}} or
\code{\link{import_all_interactions}}}
}
\value{
An igraph object
}
\description{
Transforms the interactions data frame to an igraph graph object.
}
\examples{
\donttest{
interactions = import_omnipath_interactions(resources=c('SignaLink3'))
OPI_g = interaction_graph(interactions)
}

}
\seealso{
\code{\link{import_omnipath_interactions}},
    \code{\link{import_pathwayextra_interactions}},
    \code{\link{import_kinaseextra_interactions}},
    \code{\link{import_ligrecextra_interactions}},
    \code{\link{import_dorothea_interactions}},
    \code{\link{import_mirnatarget_interactions}} or
    \code{\link{import_all_interactions}}
}
