% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercell.R
\name{intercell_categories}
\alias{intercell_categories}
\alias{get_intercell_categories}
\title{Categories in the intercell database of OmniPath}
\usage{
intercell_categories()
}
\value{
character vector with the different intercell categories
}
\description{
Retrieves a list of categories from \url{https://omnipath.org/intercell}.
}
\examples{
intercell_categories()

}
\seealso{
\itemize{
    \item{\code{\link{intercell}}}
    \item{\code{\link{intercell_generic_categories}}}
    \item{\code{\link{intercell_summary}}}
}
}
