% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{intercell_consensus_filter}
\alias{intercell_consensus_filter}
\title{Quality filter for intercell annotations}
\usage{
intercell_consensus_filter(data, percentile = NULL)
}
\arguments{
\item{data}{A data frame with intercell annotations, as provided by
\code{\link{import_omnipath_intercell}}.}

\item{percentile}{Numeric: a percentile cut off for the consensus score
of composite categories. The consensus score is the number of
resources supporting the classification of an entity into a category
based on combined information of many resources. Here you can apply
a cut-off, keeping only the annotations supported by a higher number
of resources than a certain percentile of each category. If
\code{NULL} no filtering will be performed. The value is either in the
0-1 range, or will be divided by 100 if graeter than 1. The
percentiles will be calculated against the generic composite
categories and then will be applied to their resource specific
annotations and specific child categories.}
}
\value{
The data frame in \code{data} filtered by the consensus scores.
}
\description{
Quality filter for intercell annotations
}
\examples{
intercell <- import_omnipath_intercell(parent = c('ligand', 'receptor'))
nrow(intercell)
# [1] 50174
intercell_q50 <- intercell_consensus_filter(intercell, 50)
nrow(intercell_q50)
# [1] 41562

}
