% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggrest.R
\name{kegg_request}
\alias{kegg_request}
\title{Perform a KEGG REST API request}
\usage{
kegg_request(operation, ...)
}
\arguments{
\item{operation}{Character: one of the KEGG REST API operations.}

\item{...}{Arguments for the API operation, as defined in the templates
available by \code{\link{kegg_api_templates}} and in the page
\url{https://www.kegg.jp/kegg/rest/keggapi.html}.}
}
\value{
List or data frame: the data retrieved from the KEGG REST API.
}
\description{
Perform a KEGG REST API request
}
\examples{
kegg_request("conv", "compound", "pubchem")

}
