% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmanager.R
\name{load_db}
\alias{load_db}
\title{Load a built in database}
\usage{
load_db(key, param = list())
}
\arguments{
\item{key}{Character: the key of the database to load. For a list of
available keys see \code{\link{omnipath_show_db}}.}

\item{param}{List: override the defaults or pass further parameters to
the database loader function. See the loader functions and their
default parameters in \code{\link{omnipath_show_db}}.}
}
\description{
Load a built in database
}
\examples{
load_db('gene_ontology_slim')

}
\seealso{
\code{\link{omnipath_show_db}}.
}
