% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatlas.R
\name{metabolic_atlas_list_models}
\alias{metabolic_atlas_list_models}
\title{List of original GEMs in Metabolic Atlas}
\usage{
metabolic_atlas_list_models(integrated = FALSE)
}
\arguments{
\item{integrated}{Logical: list the integrated (standard) GEMs. These are
available by a separate API in Metabolic Atlas. There are 7 integrated
GEMs available by the API, and 23 standard GEMs listed in Metabolic
Atlas in total. In the repository a total of 360 models are available.}
}
\value{
A data frame (tibble) of GEMs.
}
\description{
Metabolic Atlas is a repository of genome scale models of metabolism (GEMs)
for various organisms, tissues and conditions. This function returns a list
of the available GEMs. Read more at \url{https://metabolicatlas.org}.
}
\examples{
metabolic_atlas_models()

}
