% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{nichenet_signaling_network_cpdb}
\alias{nichenet_signaling_network_cpdb}
\title{Builds signaling network prior knowledge for NicheNet using ConsensusPathDB
(CPDB)}
\usage{
nichenet_signaling_network_cpdb(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A network data frame (tibble) with signaling interactions
    suitable for use with NicheNet.
}
\description{
Builds signaling network prior knowledge for NicheNet using ConsensusPathDB
(CPDB)
}
\examples{
# use some parameters stricter than default:
cpdb_signaling_network <- nichenet_signaling_network_cpdb(
    complex_max_size = 2,
    min_score = .99
)

}
\seealso{
\code{\link{nichenet_signaling_network},
    \link{consensuspathdb_download}}
}
