% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{omnipath_log}
\alias{omnipath_log}
\title{Browse the current OmnipathR log file}
\usage{
omnipath_log()
}
\description{
Browse the current OmnipathR log file
}
\examples{
\donttest{
omnipath_log()
# then you can browse the log file, and exit with `q`
}

}
\seealso{
\code{\link{omnipath_logfile}}
}
