% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{omnipath_logfile}
\alias{omnipath_logfile}
\alias{logfile}
\title{Path to the current OmnipathR log file}
\usage{
omnipath_logfile()

logfile(pkg = "OmnipathR")
}
\arguments{
\item{pkg}{Character: name of a package.}
}
\value{
Character: path to the current logfile, or \code{NULL} if no
    logfile is available.
}
\description{
Path to the current OmnipathR log file

Path to the current logfile of a package
}
\examples{
omnipath_logfile()
# [1] "/home/denes/omnipathr/omnipathr-log/omnipathr-20210309-1642.log"

}
\seealso{
\code{\link{omnipath_log}}
}
