% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{omnipath_reset_config}
\alias{omnipath_reset_config}
\title{Restores the built-in default values of all config parameters}
\usage{
omnipath_reset_config(save = NULL)
}
\arguments{
\item{save}{If a path, the restored config will be also saved
to this file. If TRUE, the config will be saved to the current default
config path (see \code{\link{omnipath_get_config_path}}).}
}
\value{
The config as a list.
}
\description{
Restores the built-in default values of all config parameters
}
\examples{
\donttest{
# restore the defaults and write them to the default config file:
omnipath_reset_config()
omnipath_save_config()
}

}
\seealso{
\code{\link{omnipath_load_config}, \link{omnipath_save_config}}
}
