% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmanager.R
\name{omnipath_show_db}
\alias{omnipath_show_db}
\title{Built in database definitions}
\usage{
omnipath_show_db()
}
\value{
A data frame with the built in database definitions.
}
\description{
Databases are resources which might be costly to load but can be used many
times by functions which usually automatically load and retrieve them from
the database manager. Each database has a lifetime and will be unloaded
automatically upon expiry.
}
\examples{
database_definitions <- omnipath_show_db()
database_definitions
# # A tibble: 14 x 10
#    name       last_used           lifetime package  loader    loader_p.
#    <chr>      <dttm>                 <dbl> <chr>    <chr>     <list>
#  1 Gene Onto. 2021-04-04 20:19:15      300 Omnipat. go_ontol. <named l.
#  2 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
#  3 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
#  4 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
#  5 Gene Onto. NA                       300 Omnipat. go_ontol. <named l.
# ... (truncated)
# # . with 4 more variables: latest_param <list>, loaded <lgl>, db <list>,
# #   key <chr>

}
