% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{pivot_annotations}
\alias{pivot_annotations}
\title{Converts annotation tables to a wide format}
\usage{
pivot_annotations(annotations)
}
\arguments{
\item{annotations}{A data frame of annotations downloaded from the
OmniPath web service.}
}
\value{
A wide format tibble if the provided data contains annotations
from one resource, otherwise a list of wide format tibbles.
}
\description{
Use this method to reconstitute the annotation tables into the format of
the original resources. With the `wide=TRUE` option
\code{import_omnipath_annotations} applies this function to the downloaded
data.
}
