% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{print_interactions}
\alias{print_interactions}
\title{Print OmniPath interactions}
\usage{
print_interactions(interDF, writeRefs = FALSE)
}
\arguments{
\item{interDF}{data.frame with the interactions generated by any of the
following functions: \itemize{
\item{\code{\link{import_omnipath_enzsub}}}
\item{\code{\link{import_omnipath_interactions}}}
\item{\code{\link{import_pathwayextra_interactions}}}
\item{\code{\link{import_kinaseextra_interactions}}}
\item{\code{\link{import_ligrecextra_interactions}}}
\item{\code{\link{import_post_translational_interactions}}}
\item{\code{\link{import_dorothea_interactions}}}
\item{\code{\link{import_tf_target_interactions}}}
\item{\code{\link{import_transcriptional_interactions}}}
\item{\code{\link{import_mirnatarget_interactions}}}
\item{\code{\link{import_all_interactions}}}}}

\item{writeRefs}{[FALSE] writes also the PubMed IDs if available}
}
\value{
Returns `NULL`.
}
\description{
Prints the interactions or enzyme-substrate relationships in a nice format.
}
\examples{
enzsub <- import_omnipath_enzsub()
print_interactions(head(enzsub))
print_interactions(tail(enzsub), writeRefs = TRUE)
print_interactions(
    dplyr::filter(
        enzsub,
        enzyme_genesymbol == 'MAP2K1',
        substrate_genesymbol == 'MAPK3'
    )
)

signor <- import_omnipath_interactions(resources = 'SIGNOR')
print_interactions(head(signor))
#            source interaction            target n_resources
# 6 MAPK14 (Q16539)  ==( + )==> MAPKAPK2 (P49137)          23
# 4  TRPM7 (Q96QT4)  ==( + )==>    ANXA1 (P04083)          10
# 1  PRKG1 (Q13976)  ==( - )==>    TRPC3 (Q13507)           8
# 2  PTPN1 (P18031)  ==( - )==>    TRPV6 (Q9H1D0)           6
# 5 PRKACA (P17612)  ==( - )==>   MCOLN1 (Q9GZU1)           6
# 3  RACK1 (P63244)  ==( - )==>    TRPM6 (Q9BX84)           2
}
