% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{print_interactions}
\alias{print_interactions}
\title{Print OmniPath interactions}
\usage{
print_interactions(interDF, writeRefs = FALSE)
}
\arguments{
\item{interDF}{data.frame with the interactions generated by any of the
following functions: \code{\link{import_omnipath_enzsub}},
\code{\link{import_omnipath_interactions}},
\code{\link{import_pathwayextra_interactions}},
\code{\link{import_kinaseextra_interactions}},
\code{\link{import_ligrecextra_interactions}},
\code{\link{import_dorothea_interactions}},
\code{\link{import_mirnatarget_interactions}} or
\code{\link{import_all_interactions}}}

\item{writeRefs}{[FALSE] writes also the PubMed IDs if available}
}
\description{
Prints the interactions or enzyme-substrate relationships in a nice format.
}
\examples{
ptms = import_omnipath_enzsub()
print_interactions(head(ptms))
print_interactions(tail(ptms), writeRefs = TRUE)
print_interactions(
    dplyr::filter(
        ptms,
        enzyme_genesymbol == 'MAP2K1',
        substrate_genesymbol == 'MAPK3'
    )
)
}
