% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{print_path_es}
\alias{print_path_es}
\alias{printPath_es}
\title{print network paths given by edge sequence}
\usage{
print_path_es(edgeSeq, G)

printPath_es(...)
}
\arguments{
\item{edgeSeq}{edge sequence}

\item{G}{igraph object (from ptms or any interaction dataset)}

\item{...}{Passed to \code{print_path_es}.}
}
\value{
Interactions displayed in a nice format
}
\description{
Prints the interactions in the path in a nice format.
}
\examples{
interactions = import_omnipath_interactions(resources=c("SignaLink3"))
OPI_g = interaction_graph(interactions = interactions )
print_path_es(shortest_paths(OPI_g,from = "TYRO3",to = "STAT3",
    output = 'epath')$epath[[1]],OPI_g)
}
\seealso{
\code{\link{print_path_vs}}
}
