% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_interactions.R
\name{print_path_vs}
\alias{print_path_vs}
\alias{printPath_vs}
\title{print networks paths given by node sequence}
\usage{
print_path_vs(nodeSeq, G)

printPath_vs(...)
}
\arguments{
\item{nodeSeq}{node sequence}

\item{G}{igraph object (from ptms or interactions)}

\item{...}{Passed to \code{print_path_vs}.}
}
\value{
Interactions displayed in a nice format
}
\description{
Prints the interactions in the path in a nice format.
}
\examples{
interactions = import_omnipath_interactions(resources=c("SignaLink3"))
OPI_g = interaction_graph(interactions = interactions )
print_path_vs(
    all_shortest_paths(
        OPI_g,
        from = "TYRO3",
        to = "STAT3"
    )$vpath,
    OPI_g
)
ptms = import_omnipath_enzsub(resources=c("PhosphoSite", "SIGNOR"))
ptms_g = ptms_graph(ptms)
print_path_vs(
    all_shortest_paths(
        ptms_g,
        from = "SRC",
        to = "STAT1"
    )$res,
    ptms_g
)
}
\seealso{
\code{\link{print_path_es}}
}
