% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe2graph.R
\name{ptms_graph}
\alias{ptms_graph}
\title{PTMS graph}
\usage{
ptms_graph(ptms)
}
\arguments{
\item{interactions}{data.frame created by \code{\link{import_Omnipath_PTMS}}}
}
\value{
igraph object
}
\description{
transforms the ptmds interactions data.frame to igraph object
}
\examples{
ptms = import_Omnipath_PTMS(filter_databases=c("PhosphoSite", "Signor"))
ptms_g = ptms_graph(ptms = ptms )
}
