% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{query_info}
\alias{query_info}
\title{OmniPath query parameters}
\usage{
query_info(query_type)
}
\arguments{
\item{query_type}{Character: interactions, annotations, complexes, enz_sub
or intercell.}
}
\value{
A named list with the parameter names and their possible values.
}
\description{
All parameter names and their possible values for a query type. Note:
parameters with `NULL` values have too many possible values to list
them.
}
\examples{
ia_param <- query_info('interactions')
ia_param$datasets[1:5]
# [1] "dorothea"    "kinaseextra" "ligrecextra" "lncrna_mrna" "mirnatarget"

}
