% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnetwork.R
\name{regnetwork_download}
\alias{regnetwork_download}
\title{Interactions from RegNetwork}
\usage{
regnetwork_download(organism = "human")
}
\arguments{
\item{organism}{Character: either human or mouse.}
}
\value{
Data frame with interactions
}
\description{
Downloads transcriptional and post-transcriptional regulatory interactions
from the RegNetwork database (http://www.regnetworkweb.org/).
}
\examples{
\donttest{
regn_interactions <- regnetwork_download()
}

}
