% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnipath.R
\name{resources_colname}
\alias{resources_colname}
\title{Name of the column with the resources}
\usage{
resources_colname(data)
}
\value{
Character: the name of the column, if any of the column names
    matches.
}
\description{
Unfortunately the column title is different across the various
query types in the OmniPath web service, so we need to guess.
}
\examples{
co <- import_omnipath_complexes()
resources_colname()

}
