% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_omnipath}
\alias{search_omnipath}
\title{Search in OmniPath}
\usage{
search_omnipath(..., rebuild_index = FALSE)
}
\arguments{
\item{...}{Ignored.

Search for keywords in the resource and attribute names and metadata.}

\item{rebuild_index}{Logical: rebuild the index even if it could be loaded
from the cache.}
}
\description{
Search in OmniPath
}
