% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_actions}
\alias{stitch_actions}
\title{Retrieve the STITCH actions dataset}
\usage{
stitch_actions(organism = "human", prefixes = FALSE)
}
\arguments{
\item{organism}{Character or integer: name or NCBI Taxonomy ID of an
organism. STICH supports many organisms, please refer to their web site
at \url{https://stitch.embl.de/}.}
}
\value{
Data frame of STITCH actions.
}
\description{
Retrieve the STITCH actions dataset
}
\examples{
sta <- stitch_actions(organism = 'mouse')

}
\seealso{
\itemize{
    \item{\code{\link{stitch_actions}}}
    \item{\code{\link{stitch_links}}}
    \item{\code{\link{stitch_network}}}
}
}
