% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch_links}
\alias{stitch_links}
\title{Retrieve the STITCH links dataset}
\usage{
stitch_links(organism = "human", prefixes = FALSE)
}
\arguments{
\item{organism}{Character or integer: name or NCBI Taxonomy ID of an
organism. STICH supports many organisms, please refer to their web site
at \url{https://stitch.embl.de/}.}
}
\value{
Data frame: organism specific STITCH links dataset.
}
\description{
Retrieve the STITCH links dataset
}
\examples{
stl <- stitch_links()

}
\seealso{
\itemize{
    \item{\code{\link{stitch_actions}}}
    \item{\code{\link{stitch_links}}}
    \item{\code{\link{stitch_network}}}
}
}
