% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{translate_ids}
\alias{translate_ids}
\title{Translates a column of identifiers in a data frame by creating another
column with the target identifiers}
\usage{
translate_ids(
  d,
  from_col,
  to_col,
  from,
  to,
  uploadlists = TRUE,
  keep_untranslated = TRUE,
  ...
)
}
\arguments{
\item{d}{A data frame}

\item{from_col}{Name of an existing column in the data frame containing
the identifiers to be translated}

\item{to_col}{Name for a new column which will contain the target
identifiers}

\item{from}{Identifier type for `from_col`. See Details for possible
values.}

\item{to}{Identifier type for `to_col`. See Details for possible values.}

\item{uploadlists}{Whether to use the `uploadlists` service from UniProt
or the plain query interface (implemented in
\code{\link{uniprot_full_id_mapping_table}} in this package).}

\item{keep_untranslated}{Keep the records where the source identifier
could not be translated. At these records the target identifier will
be NA.}

\item{...}{Passed to \code{\link{uniprot_full_id_mapping_table}}.}
}
\value{
The data frame `d` a new column added with the translated IDs.
}
\description{
Translates a column of identifiers in a data frame by creating another
column with the target identifiers
}
\details{
This function uses the uploadlists service of UniProt to obtain identifier
translation tables. The possible values for `from` and `to` are the
identifier type abbreviations used in the UniProt API, please refer to
the table here: https://www.uniprot.org/help/api_idmapping
The mapping between identifiers can be ambiguous. In this case one row
in the original data frame yields multiple rows in the returned data
frame.
}
\examples{
d <- data.frame(uniprot_id = c('P00533', 'Q9ULV1', 'P43897', 'Q9Y2P5'))
d <- translate_ids(d, uniprot_id, genesymbol, 'ID', 'GENENAME')
d
#   uniprot_id  genesymbol
# 1     P00533  EGFR_HUMAN
# 2     Q9ULV1  FZD4_HUMAN
# 3     P43897  EFTS_HUMAN
# 4     Q9Y2P5 S27A5_HUMAN

}
\seealso{
\code{\link{uniprot_id_mapping_table}}
}
