% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{translate_ids_multi}
\alias{translate_ids_multi}
\title{Translate gene, protein and small molecule identifiers from multiple columns}
\usage{
translate_ids_multi(
  d,
  ...,
  suffixes = NULL,
  suffix_sep = "_",
  uploadlists = FALSE,
  ensembl = FALSE,
  hmdb = FALSE,
  chalmers = FALSE,
  entity_type = NULL,
  keep_untranslated = TRUE,
  return_df = FALSE,
  organism = 9606,
  reviewed = TRUE
)
}
\description{
Especially when translating network interactions, where two ID columns exist
(source and target), it is convenient to call the same ID translation on
multiple columns. The \code{\link{translate_ids}} function is already able
to translate to multiple ID types in one call, but is able to work only from
one source column. Here too, multiple target IDs are supported. The source
columns can be listed explicitely, or they might share a common stem, in
this case the first element of \code{...} will be used as stem, and the
column names will be created by adding the \code{suffixes}. The
\code{suffixes} are also used to name the target columns. If no
\code{suffixes} are provided, the name of the source columns will be added
to the name of the target columns. ID types can be defined the same way as
for \code{\link{translate_ids}}. The only limitation is that, if the source
columns are provided as stem+suffixes, they must be the same ID type.
}
