% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_full_id_mapping_table}
\alias{uniprot_full_id_mapping_table}
\title{Creates an ID translation table from UniProt data}
\usage{
uniprot_full_id_mapping_table(
  to,
  from = "id",
  reviewed = TRUE,
  organism = 9606
)
}
\arguments{
\item{to}{Target ID type. See Details for possible values.}

\item{from}{Source ID type. See Details for possible values.}

\item{reviewed}{Retrieve only reviewed (TRUE), only unreviewed (FALSE) or
both (NULL).}

\item{organism}{Integer, NCBI Taxonomy ID of the organism (by default
9606 for human).}
}
\description{
Creates an ID translation table from UniProt data
}
\details{
For both source and target ID type, this function accepts column codes
used by UniProt and some simple shortcuts defined here. For the UniProt
codes please refer to
https://www.uniprot.org/help/uniprotkb%5Fcolumn%5Fnames
The shortcuts are entrez, genesymbol, genesymbol_syn (synonym gene
symbols), hgnc, embl, refseqp (RefSeq protein), enst (Ensembl transcript),
uniprot_entry (UniProtKB AC, e.g. EGFR_HUMAN), protein_name (full name of
the protein), uniprot (UniProtKB ID, e.g. P00533).
}
\examples{
uniprot_entrez <- uniprot_full_id_mapping_table(to = 'entrez')
}
